% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{add_peak_list}
\alias{add_peak_list}
\title{SCZip - add_peak_list}
\usage{
SCZip$add_peak_list(peak_list_data)
}
\arguments{
\item{peak_list_data}{a \code{\link[=Peaks]{Peaks()}} object}
}
\description{
adds a peak list to the \code{SCZip}
}
