% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeMS.R
\name{add_scans_to_file}
\alias{add_scans_to_file}
\title{add scans to file}
\usage{
add_scans_to_file(zip_file, out_file = zip_file, exclude_noise = TRUE)
}
\arguments{
\item{zip_file}{the zip file to work with}

\item{out_file}{if desired, a new file name to generate}
}
\description{
adds the jsonified scan peaks to an already existing zip file
}
