% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{add_to_zip}
\alias{add_to_zip}
\title{add to zip file}
\usage{
add_to_zip(object, filename, zip_file)
}
\arguments{
\item{object}{the object to write}

\item{filename}{the file that it should be}

\item{zip_file}{the zip file to add to}
}
\description{
given an output object, filename, and zip file, write the output object
to the file, and then add to the zip file
}
\details{
a directory created by \code{tempdir} is used to hold the file,
which is then added to the zip file.
}

