% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeMS.R
\docType{data}
\name{AnalyzeMS}
\alias{AnalyzeMS}
\title{analyze ftms mass-spec data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnalyzeMS
}
\description{
This class allows you to analyze mass spec data, and controls the execution
of reading in the mass
}
\keyword{datasets}

