% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_normalization.R
\name{anova_matrix_test}
\alias{anova_matrix_test}
\title{anova matrix}
\usage{
anova_matrix_test(intensity, groups)
}
\arguments{
\item{intensity}{the intensity matrix, features as rows and samples as columns}

\item{groups}{vector of factors, named the same as the columns of the intensity matrix}
}
\value{
data.frame
}
\description{
For an entire matrix, perform ANOVA across defined groups for each row. The
groups should be a factor vector, and have the same names as the columns
of the intensity matrix.
}
