% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_normalization.R
\name{anova_test}
\alias{anova_test}
\title{anova test}
\usage{
anova_test(intensity, groups)
}
\arguments{
\item{intensity}{the intensity values}

\item{groups}{vector giving the groups of the samples}
}
\value{
data.frame
}
\description{
perform an ANOVA across groups and return a tidy data.frame of the statistical
results
}
