% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{area_sum_points}
\alias{area_sum_points}
\title{area sum}
\usage{
area_sum_points(peak_mz, peak_intensity, zero_value = 0)
}
\arguments{
\item{peak_mz}{the mz in the peak}

\item{peak_intensity}{the peak intensities}

\item{zero_value}{what value actually represents zero}
}
\value{
numeric
}
\description{
calculate the area based on summing the points
}

