% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{basic_peak_center_intensity}
\alias{basic_peak_center_intensity}
\title{basic peak center}
\usage{
basic_peak_center_intensity(mz, intensity)
}
\arguments{
\item{mz}{the mz values}

\item{intensity}{the intensity values}
}
\value{
numeric
}
\description{
given the peak, finds center based on the mean, and intensity by averaging
between the two middle points.
}

