% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{generate bootstrap samples}
\usage{
bootstrap_samples(indices, n_bootstrap = 1000, n_sample = NULL)
}
\arguments{
\item{indices}{the indices to take the bootstrap sample from}

\item{n_bootstrap}{how many bootstrap samples to generate? (default is 1000)}

\item{n_sample}{how many items should be in each sample? See \strong{Details}.}
}
\value{
list of bootstrap sampled indices
}
\description{
Generates \code{n_bootstrap} samples, each with \code{n_sample} entries, from the provided
\code{indices}. See \strong{Details} for more information.
}
\details{
A \emph{bootstrap} sample is a sample where the entries have been sampled
\strong{\href{https://en.wikipedia.org/wiki/Bootstrapping_(statistics)}{with replacement}}.
In general, a \emph{bootstrap} sample has the same number of entries as the original
sample. However, there may be cases where \emph{upsampling} to specific number
may be useful. Therefore, the parameter \code{n_sample} is provided to enable
\emph{upsampling} the original indices to a larger number. If \code{n_sample} is NULL,
then no \emph{upsampling} will be done.
}
