% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{calc_point_slopes}
\alias{calc_point_slopes}
\title{calculate subsequent slopes}
\usage{
calc_point_slopes(mz, intensity)
}
\arguments{
\item{mz}{the mz values}

\item{intensity}{the intensity values}
}
\value{
numeric
}
\description{
for a set of mz - intensity point pairs, get the point-to-point slopes
}
