% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{calculate_density}
\alias{calculate_density}
\title{peak density}
\usage{
calculate_density(peak_data, use_range = NULL, window = 1, delta = 0.1)
}
\arguments{
\item{peak_data}{numeric vector of m/z values representing peaks}

\item{use_range}{the range of m/z's to generate windows}

\item{window}{how big an m/z window to calculate density}

\item{delta}{how much to slide the window}
}
\value{
data.frame
}
\description{
calculates peak density in a sliding window of m/z
}
