% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{calculate_mz_offsets}
\alias{calculate_mz_offsets}
\title{calculate M/Z offsets}
\usage{
calculate_mz_offsets(master_list, multi_peaklist)
}
\arguments{
\item{master_list}{A master sample or peak list object}

\item{multi_peaklist}{A multi sample or multi scan peak list object}
}
\value{
data.frame
}
\description{
It is useful to be able to compare the original M/Z values for a scan or sample,
and the offset adjusted M/Z values after correspondence. This function facilitates
that calculation.
}
