% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_normalization.R
\name{calculate_rsd}
\alias{calculate_rsd}
\title{rsd calculation}
\usage{
calculate_rsd(intensity, na.rm = TRUE)
}
\arguments{
\item{intensity}{the vector of intensities}
}
\value{
double
}
\description{
calculate the relative standard deviation across all samples. Note that by default
NA and Inf values are removed, so they do not need to be filtered out first.
}
