% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{characterize_peaks}
\alias{characterize_peaks}
\title{characterize peaks from points and picked peaks}
\usage{
characterize_peaks(peak_region, calculate_peak_area = FALSE)
}
\arguments{
\item{peak_region}{the PeakRegion object to work on}
}
\value{
list
}
\description{
characterize peaks from points and picked peaks
}
