% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\docType{data}
\name{CharacterizeMSPeaks}
\alias{CharacterizeMSPeaks}
\title{analyze ftms mass-spec data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CharacterizeMSPeaks
}
\description{
This class allows you to analyze mass spec data, and controls the execution
of reading in the mass
}
\keyword{datasets}
