% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{check_for_zip}
\alias{check_for_zip}
\title{check for zip}
\usage{
check_for_zip()
}
\description{
FTMS.peakCharacterization uses zip files to gather all the pieces of results together,
including the original mzML, binary data file, and JSON files. R can be compiled
on systems where there is no zip function installed.
When it is not installed, it generally creates a permission issue when \code{system2}
tries to call a non-existent command.
This function checks for zip, and warns the user if it can't be found on package load.
If \code{zip} is installed, you may need to set it using \code{Sys.setenv(R_ZIPCMD = 'zip_function')}.
}
