% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{check_frequency_r2}
\alias{check_frequency_r2}
\title{check r2}
\usage{
check_frequency_r2(mz_frequency_list)
}
\arguments{
\item{mz_frequency_list}{the list of predicted frequency data.frames}
}
\description{
check r2
}
