% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{check_model_sd}
\alias{check_model_sd}
\title{check_model_sd}
\usage{
check_model_sd(predict_function, model_1, model_2, reject_frac = 0.1)
}
\value{
logical
}
\description{
Checks the fraction of predictions in a second model that are larger than
than the same predictions in the first model, and returns TRUE if that fraction
is smaller than the defined cutoff.
}
