% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{check_zip_file}
\alias{check_zip_file}
\title{check zip file}
\usage{
check_zip_file(zip_file)
}
\arguments{
\item{zip_file}{the zip file to be checked}
}
\value{
character
}
\description{
checks that the zip file has the basic contents it should have, and that
files listed in the metadata actually exist.
}

