% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{choose_peak_points_area}
\alias{choose_peak_points_area}
\title{choose the peak}
\usage{
choose_peak_points_area(area_results, min_area = 0.1)
}
\arguments{
\item{area_results}{vector of information from \code{test_peak_area}}

\item{min_area}{the minimum area to consider}
}
\value{
numeric
}
\description{
given a matrix that defines the possible sets of points for a parabolic,
choose the longest set with an r-squared fit above a certain criteria
}
