% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{choose_peak_rsq}
\alias{choose_peak_rsq}
\title{choose peak rsq}
\usage{
choose_peak_rsq(rsq_results, min_rsq = 0.98)
}
\arguments{
\item{rsq_results}{matrix of results from \code{test_peak_rsq}}

\item{min_rsq}{minimum cutoff for the fit to use}
}
\value{
numeric
}
\description{
choose peak rsq
}

