% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRaw.R
\name{choose_single_frequency_model_default}
\alias{choose_single_frequency_model_default}
\title{default single model}
\usage{
choose_single_frequency_model_default(sc_raw)
}
\arguments{
\item{sc_raw}{the sc_raw object}
}
\value{
SCRaw
}
\description{
default single model
}
\details{
This is the default function to choose a single frequency and mz model.
It takes the scan_info after filtering scans, and calculates the median of the
square root terms, and chooses the one closest to the median value.

Please examine this function and write your own if needed.
}
