% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMS.R
\name{choose_single_model_default}
\alias{choose_single_model_default}
\title{default single model}
\usage{
choose_single_model_default(raw_ms)
}
\arguments{
\item{raw_ms}{the raw_ms object}
}
\value{
RawMS
}
\description{
default single model
}
\details{
This is the default function to choose a single frequency and mz model.
It takes the ms_info after filtering scans, and calculates the median of the
square root terms, and chooses the one closest to the median value.

Please examine this function and write your own as needed.
}
