% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{collapse_correspondent_peaks}
\alias{collapse_correspondent_peaks}
\title{collapse correspondent peaks}
\usage{
collapse_correspondent_peaks(mpl)
}
\arguments{
\item{mpl}{the \code{MasterPeakList} object}
}
\description{
Given a \code{MasterPeakList} object, examines the m/z values for subsequent
peaks to find peaks within tolerances, and checks to see if two peaks should
be collapsed based on some simple heuristics.
}
