% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{compare_master_peak_lists}
\alias{compare_master_peak_lists}
\title{compare two MasterPeakList objects}
\usage{
compare_master_peak_lists(mpl_1, mpl_2, compare_list = c("master", "scan",
  "scan_intensity", "scan_mz"))
}
\arguments{
\item{mpl_1}{the first one}

\item{mpl_2}{the second one}

\item{compare_list}{which pieces to compare}
}
\description{
Given two MasterPeakList objects, make a determination as to whether they
are the same or different
}

