% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{compare_model_predictions}
\alias{compare_model_predictions}
\title{compare two sets of model predictions}
\usage{
compare_model_predictions(model_pred_1, model_pred_2, pred_column = "y")
}
\arguments{
\item{model_pred_1}{the first one}

\item{model_pred_2}{the second}

\item{pred_column}{which column in the data.frame to use for comparison}
}
\value{
data.frame
}
\description{
Given two different model prediction objects, use the L1 Norm (maximum absolute difference)
to decide if they are the same or not.
}
