% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{compare_mpl_to_list}
\alias{compare_mpl_to_list}
\title{compare MasterPeakList to others}
\usage{
compare_mpl_to_list(mpl_check, mpl_list, min_check = 3,
  exclude_check = NULL)
}
\arguments{
\item{mpl_check}{the list of things}

\item{mpl_list}{the one to check}

\item{min_check}{what value is a minimum valid check}

\item{exclude_check}{is there anything to exclude??}
}
\description{
Given one MasterPeakList and a bunch of others, determine if there is a match
to any of the previous ones
}
