% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{compare_object_to_list_diff}
\alias{compare_object_to_list_diff}
\title{compare objects to others 2}
\usage{
compare_object_to_list_diff(object_check, object_list, min_check = 3,
  exclude_check = NULL, check_function)
}
\arguments{
\item{object_check}{the list of things}

\item{object_list}{the one to check}

\item{min_check}{what value is a minimum valid check}

\item{exclude_check}{is there anything to exclude??}

\item{check_function}{what function should be used to compare the objects?}
}
\description{
Given one object and a list of others, determine if there is a match
to any of the previous ones
}
\details{
this function allows one to check that an object "matches" any of the
  objects in a provided list. How an object "matches" is determined by the
  \emph{check_function}, it should merely return TRUE or FALSE
}
