% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{compare_slopes}
\alias{compare_slopes}
\title{compare slopes}
\usage{
compare_slopes(values)
}
\arguments{
\item{values}{the values to work with}
}
\value{
logical
}
\description{
check if subsequent differences show a minima with at least two increasing
points after the minima
}
