% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{convert_mz_frequency}
\alias{convert_mz_frequency}
\title{Convert M/Z to Frequency}
\usage{
convert_mz_frequency(mz_data, valid_range = c(0.49, 0.51),
  keep_all = FALSE)
}
\arguments{
\item{mz_data}{a data.frame with \code{mz}}

\item{valid_range}{points with a frequency difference in this range can be used for modeling}

\item{keep_all}{keep \strong{all} the variables generated, or just the original + \strong{frequency}?}
}
\value{
data.frame
}
\description{
Given a data.frame of m/z, generate frequency values for the data.
}
\details{
The \strong{M/Z} values from FTMS data do not have constant spacing between them.
This produces challenges in working with ranged intervals and windows. The
solution for FTMS data then is to convert them to \strong{frequency} space. This
is done by:
\itemize{
\item taking subsequent M/Z points
\item averaging their M/Z
\item taking the difference to get an \code{offset} value
\item dividing averaged M/Z by offset to generate \strong{frequency}
\item taking subsequent differences of frequency points
\item keep points with a difference in the supplied range as valid for modeling
}

After deciding on the valid points for modeling, each point gets an interpolated
frequency value using the two averaged points to the left and right in M/Z.
}
\seealso{
mz_scans_to_frequency
}
