% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_sd_intensity_corrections.R
\name{correct_mean}
\alias{correct_mean}
\title{correct peak mean}
\usage{
correct_mean(observed_mean, corrected_sd, fraction)
}
\arguments{
\item{observed_mean}{the observed mean}

\item{corrected_sd}{a corrected sd, generated by \code{correct_variance}}

\item{fraction}{the fraction of total observations}
}
\value{
corrected mean
}
\description{
Given a corrected SD, corrects the mean assuming that it is the result of
a truncated normal distribution.
}
\references{
https://en.wikipedia.org/wiki/Truncated_normal_distribution
}
\seealso{
\code{\link[=correct_peak]{correct_peak()}} \code{\link[=correct_variance]{correct_variance()}}
}
