% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_sd_intensity_corrections.R
\name{correct_peak}
\alias{correct_peak}
\title{correct peak sd and mean}
\usage{
correct_peak(observed_mean, observed_sd, n_observed, n_should_observe)
}
\arguments{
\item{observed_mean}{the observed mean}

\item{observed_sd}{the observed sd}

\item{n_observed}{how many observations went into this mean}

\item{n_should_observe}{how many observations \emph{should} there have been?}
}
\value{
data.frame, with corrected mean and sd
}
\description{
Assuming that an observed mean (intensity) and sd are from a truncated normal distribution
that is truncated on one side only.
}
\references{
https://en.wikipedia.org/wiki/Truncated_normal_distribution
}
\seealso{
correct_mean correct_variance
}
