% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_sd_intensity_corrections.R
\name{correct_peak_sd_height}
\alias{correct_peak_sd_height}
\title{correct peak height and sd}
\usage{
correct_peak_sd_height(
  original_height,
  list_of_heights,
  n_observed,
  n_should_observe
)
}
\arguments{
\item{original_height}{the original height estimate to correct}

\item{list_of_heights}{the set of peak heights}

\item{n_observed}{how many were observed}

\item{n_should_observe}{how many should have been observed}
}
\value{
data.frame
}
\description{
correct peak height and sd
}
