% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{CorrespondentPeakList}
\alias{CorrespondentPeakList}
\title{CorrespondentPeakList}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CorrespondentPeakList
}
\description{
Creates a PeakList compatible object from a correspondent object
}
\keyword{datasets}
