% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMS.R
\name{count_raw_peaks}
\alias{count_raw_peaks}
\title{count raw peaks}
\usage{
count_raw_peaks(rawdata, scans)
}
\arguments{
\item{rawdata}{the raw_data bit of the RawMS object}

\item{scans}{which scans to use.}
}
\value{
numeric
}
\description{
from a RawMS object, get the scans, average them, and count the peaks.
}
