% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{create_frequency_regions}
\alias{create_frequency_regions}
\title{create frequency regions}
\usage{
create_frequency_regions(
  point_spacing = 0.5,
  frequency_range = NULL,
  n_point = 10,
  delta_point = 1,
  multiplier = 500
)
}
\arguments{
\item{point_spacing}{how far away are subsequent points.}

\item{frequency_range}{the range of frequency to use}

\item{n_point}{how many points you want to cover}

\item{delta_point}{the \emph{step} size between the beginning of each subsequent region}

\item{multiplier}{multiplier to convert from frequency to integer space}
}
\value{
IRanges
}
\description{
Given a point-point spacing and a frequency range, create IRanges based regions
of specified width. Overlapping sliding regions can be creating by specifying
a \code{region_size} bigger than \code{delta}, adjacent tiled regions can be created
by specifying a \code{region_size} == \code{delta}.
}
\details{
For Fourier-transform mass spec, points are equally spaced in
frequency space, which will lead to unequal spacing in M/Z space. Therefore,
we create regions using the point-point differences in frequency space.

What will be returned is an \code{IRanges} object, where the widths are constantly
increasing over M/Z space.
}
