% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{create_mz_regions}
\alias{create_mz_regions}
\title{create mz regions}
\usage{
create_mz_regions(mz_model, use_range = NULL, region_size = 10,
  delta = 1, point_multiplier = 20000)
}
\arguments{
\item{mz_model}{data.frame where \code{x} is M/Z, \code{y} is difference in M/Z.}

\item{use_range}{the range of M/Z to use}

\item{region_size}{how \emph{big} is each of the regions}

\item{delta}{the \emph{step} size between the beginning of each subsequent region}

\item{point_multiplier}{multiplier to convert from M/Z to integer space}
}
\value{
IRanges
}
\description{
Given an M/Z data.frame model and a range, create IRanges based regions
of specified width. Overlapping sliding regions can be creating by specifying
a \code{region_size} bigger than \code{delta}, adjacent tiled regions can be created
by specifying a \code{region_size} == \code{delta}.
}
\details{
For Fourier-transform mass spec, points are often equally spaced
frequency space, which will lead to unequal spacing in M/Z space. Therefore,
to create regions that properly cover the same number of M/Z points, the spacing
between M/Z points has to increase as M/Z space. This leads to needing the
\code{mz_model}, where \code{x} is an M/Z value, and \code{y} is the spacing between subsequent
M/Z points.

What will be returned is an \code{IRanges} object, where the widths are constantly
increasing over M/Z space.
}
