% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{default_correct_offset_function}
\alias{default_correct_offset_function}
\title{correct offsets}
\usage{
default_correct_offset_function(master_peak_list, multi_scan_peaklist,
  min_scan = 0.1)
}
\arguments{
\item{master_peak_list}{the MasterPeakList object of correspondent peaks}

\item{multi_scan_peaklist}{the MultiScansPeakList to be corrected}

\item{min_scan}{what is the minimum number of scans a peak should be in to be used for correction.}
}
\value{
list
}
\description{
Given a MasterPeakList object and the MultiScansPeakList that generated it,
correct the m/z values using offset predictions
}
