% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{default_offset_predict_function}
\alias{default_offset_predict_function}
\title{offset predict function}
\usage{
default_offset_predict_function(model, x)
}
\arguments{
\item{model}{the model to use}

\item{x}{the new values}
}
\value{
numeric
}
\description{
The offset predictor using loess
}
