% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{default_scan_filter}
\alias{default_scan_filter}
\title{default filter}
\usage{
default_scan_filter(raw_ms)
}
\arguments{
\item{raw_ms}{the RawMS object}
}
\value{
RawMS
}
\description{
The default scan filter, currently calls `scan_time_filter`. This will
be added to all `CharacterizeMSPeaks` objects. To remove it, simply provide
another function, or set it to `NULL` after object creation.
}
