% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{default_sd_fit_function}
\alias{default_sd_fit_function}
\title{default sd fit}
\usage{
default_sd_fit_function(x, y)
}
\arguments{
\item{x}{the x values, often M/Z}

\item{y}{the y values, normally differences or RMSD}
}
\value{
list
}
\description{
The default fit function for differences and RMSD
}
