% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeMS.R
\name{default_sd_predict_function}
\alias{default_sd_predict_function}
\title{default sd predict}
\usage{
default_sd_predict_function(model, x)
}
\arguments{
\item{model}{the previously generated model}

\item{x}{the new data}
}
\value{
numeric
}
\description{
The default predict function for difference and RMSD
}
