% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{define_peak_type}
\alias{define_peak_type}
\title{define peak type}
\usage{
define_peak_type(peak_data, flat_cut = 0.98)
}
\arguments{
\item{peak_data}{data.frame of mz and intensity}

\item{flat_cut}{what cutoff should be used to say a peak is "flat"?}
}
\value{
list
}
\description{
Given the peak mz and intensity, decide what type of peak it is and roughly
where the peak should be.
}
