% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{df_of_model_predictions}
\alias{df_of_model_predictions}
\title{all model predictions}
\usage{
df_of_model_predictions(predict_function, x, list_of_models)
}
\arguments{
\item{predict_function}{the function used to make the predictions}

\item{x}{the thing to make predictions with}

\item{list_of_models}{the set of models}
}
\description{
When a bunch of offset models are generated, we want to go through and create
a whole set of predictions. This enables creation of a whole lot of predictions.
}
