% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{.get_scan_polarity}
\alias{.get_scan_polarity}
\title{get_scan_mode}
\usage{
.get_scan_polarity(spectrum_list)
}
\arguments{
\item{spectrum_list}{the list of spectra}
}
\description{
takes a list from xmlToList for "run" and looks at whether all scans are positive, negative, or mixed
}
