% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{.remove_attrs}
\alias{.remove_attrs}
\title{remove attributes}
\usage{
.remove_attrs(in_list)
}
\arguments{
\item{in_list}{the list to work on}
}
\description{
removes a list entry called ".attrs" from a list, and makes them first level
partners
}
