% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{.to_data_frame}
\alias{.to_data_frame}
\title{transform to data frame}
\usage{
.to_data_frame(in_list)
}
\arguments{
\item{in_list}{the list of xml nodes to work on}
}
\description{
transform to data frame
}
