% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{every_scan_frequency}
\alias{every_scan_frequency}
\title{get frequencies for all data}
\usage{
every_scan_frequency(zip)
}
\arguments{
\item{zip}{the zip object}
}
\value{
data.frame
}
\description{
Given a zip object, gets frequency conversions for \emph{all} of the scans, not just the MS1
or filtered scans.
}
