% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{exponential_fit}
\alias{exponential_fit}
\title{exponential fit}
\usage{
exponential_fit(x, y, w = NULL, n_exp = 1, center = FALSE)
}
\arguments{
\item{x}{the x values, independent}

\item{y}{the y values, dependent}

\item{w}{weights}

\item{n_exp}{how many exponents to use}

\item{center}{whether the X-values should be centered first or not}
}
\value{
list
}
\description{
Given a set of X's and Y's, calculates the fit for y = a + bx + cx^2 + dx^n ...
}

