% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{exponential_predict}
\alias{exponential_predict}
\title{exponential predictions}
\usage{
exponential_predict(coef, x)
}
\arguments{
\item{coef}{model coefficients}

\item{x}{the new x-values}
}
\value{
numeric
}
\description{
Given a model coefficients, and X's, calculate the Y values. Note that this
function makes a very simple assumption, that the length of the coefficients
corresponds to a number of exponential terms in the model, i.e. if the
coefficients has 3 terms, then the model was \emph{Y = a + bx + cx^2}.
}

