% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_data.R
\name{extract_raw_method}
\alias{extract_raw_method}
\title{extract raw method}
\usage{
extract_raw_method(in_file, output_type = "data.frame")
}
\arguments{
\item{in_file}{The Thermo raw file to extract}

\item{output_type}{string, data.frame or json}
}
\value{
string or data.frame
}
\description{
Given a Thermo ".raw" file, attempts to extract the "method"
definition from a translated hexdump of the file.
}
