% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMzml.R
\name{filter_scans_builtin}
\alias{filter_scans_builtin}
\title{built in filter scan function}
\usage{
filter_scans_builtin(sc_mzml)
}
\arguments{
\item{sc_mzml}{the sc_mzml object}
}
\value{
SCmzml
}
\description{
built in filter scan function
}
\details{
This is the built in filtering and removing outliers function.
It is based on the Moseley groups normal samples and experience.
However, it does not reflect everyone's experience and needs.
We expect that others have different use cases and needs, and therefore
they should create their own function and use it appropriately.

Please examine this function and write your own as needed.
It \strong{must} take an \code{SCMzml} object, work on the \code{scan_info} slot,
and then create a column with the name "keep" denoting which scans to keep.
To view the current definition, you can do \code{filter_scans_builtin}
}
