% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_peaks.R
\name{find_closest_peaks}
\alias{find_closest_peaks}
\title{find closest peaks}
\usage{
find_closest_peaks(query_data, subject_data, q_name, s_name)
}
\arguments{
\item{query_data}{the query peaks \code{tbl_df}}

\item{subject_data}{the subject peaks}

\item{q_name}{identifier for query}

\item{s_name}{identifier for subject}
}
\value{
tbl_df
}
\description{
taking a \code{query_data}, go through the \code{subject_data}
and find the closest peaks in \code{query_data}.
}

