% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{find_correspondence_scans}
\alias{find_correspondence_scans}
\title{find_correspondence_scans}
\usage{
find_correspondence_scans(multi_scan_peak_list,
  peak_calc_type = "lm_weighted", max_iteration = 20,
  digital_resolution_multiplier = 0.5, rmsd_multiplier = 3,
  max_failures = 5, mz_range = c(-Inf, Inf), notify_progress = FALSE,
  noise_function = NULL, sd_fit_function = NULL,
  sd_predict_function = NULL, offset_fit_function = NULL,
  offset_predict_function = NULL, offset_correction_function = NULL,
  remove_low_ic_scans = TRUE, collapse_peaks = FALSE)
}
\arguments{
\item{multi_scan_peak_list}{the MultiScansPeakList object to work with}

\item{peak_calc_type}{what type of peaks are being used}

\item{max_iteration}{how many iterations to allow before quitting?}

\item{digital_resolution_multiplier}{multiplier on the digital resolution}

\item{rmsd_multiplier}{the multiplier on the RMSD}

\item{max_failures}{how many times can things fail before quitting}

\item{mz_range}{the range of mz's to use}

\item{notify_progress}{provide informative messages on progress}

\item{noise_function}{the noise function to use}

\item{sd_fit_function}{function to fit the RMSD}

\item{sd_predict_function}{function for predicting RMSD cutoffs}

\item{offset_fit_function}{function to fit M/Z offsets across scans (samples)}

\item{offset_predict_function}{function to predict M/Z offsets}

\item{offset_correction_function}{function for doing offset correction}

\item{remove_low_ic_scans}{should low information content scans be removed}

\item{collapse_peaks}{whether to collapse the peaks or not}
}
\value{
FindCorrespondenceScans
}
\description{
Initializes a FindCorrespondenceScans object to actually do correspondence.
}
