% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{find peaks}
\usage{
find_peaks(avg_spectra, m = 3)
}
\arguments{
\item{avg_spectra}{input spectrum}

\item{m}{how many on each side need to be lower}
}
\value{
tbl_df
}
\description{
find peaks
}

