% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{find_peaks_diff}
\alias{find_peaks_diff}
\title{find peaks}
\usage{
find_peaks_diff(avg_spectra, m = 3)
}
\arguments{
\item{avg_spectra}{input spectrum}

\item{m}{how many on each side need to be lower}
}
\value{
tbl_df
}
\description{
find peaks
}
\details{
Was copied from a cross-validated post:
http://stats.stackexchange.com/questions/22974/how-to-find-local-peaks-valleys-in-a-series-of-data
by user stas-g
}
