% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{find_signal_regions}
\alias{find_signal_regions}
\title{find signal}
\usage{
find_signal_regions(
  regions,
  point_regions_list,
  multiplier = 1.5,
  n_point_region = 2000
)
}
\arguments{
\item{regions}{the regions we want to query}

\item{point_regions_list}{the individual points}

\item{multiplier}{how much above base quantiles to use (default = 1.5)}
}
\value{
IRanges
}
\description{
Given some regions and point_regions, find the regions that actually should
contain \strong{real} data. See \emph{details} for an explanation of what is considered
\strong{real}.
}
