% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{FindCorrespondence}
\alias{FindCorrespondence}
\title{FindCorrespondence}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindCorrespondence
}
\description{
FindCorrespondence
}
\keyword{datasets}
