% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{FindCorrespondenceSamples}
\alias{FindCorrespondenceSamples}
\title{FindCorrespondenceSamples}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindCorrespondenceSamples
}
\description{
FindCorrespondenceSamples
}
\keyword{datasets}
