% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{FindCorrespondenceScans}
\alias{FindCorrespondenceScans}
\title{generate correspondent peaks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FindCorrespondenceScans
}
\description{
generate correspondent peaks
}
\keyword{datasets}

