% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{frequency_points_to_frequency_regions}
\alias{frequency_points_to_frequency_regions}
\title{frequency points to frequency_regions}
\usage{
frequency_points_to_frequency_regions(
  frequency_data,
  frequency_variable = "frequency",
  multiplier = 400
)
}
\arguments{
\item{frequency_data}{a \code{data.frame}}

\item{frequency_variable}{which column is the \code{frequency} stored in}

\item{multiplier}{value used to convert to integers}
}
\description{
Given a set of frequency points in a data.frame, create IRanges based point "regions"
of width 1, using the \code{multiplier} to convert from a floating point double
to an \code{integer}
}
