% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{generate_scan_outlier_filter}
\alias{generate_scan_outlier_filter}
\title{Generate a scan filtering function}
\usage{
generate_scan_outlier_filter(rtime = NA, y.freq = NA)
}
\arguments{
\item{rtime}{retention time limits of scans to keep (NA)}

\item{y.freq}{y-frequency coefficient limits of scans to keep (NA)

Creates a new function that accesses the \code{scan_info} slot of
an \code{SCMzml} object, filters the scans by their retention-time and
y-frequency coefficients, tests for outliers in the y-frequency
coefficients, and denotes which scans will be kept for further
processing.

\code{NA} means no filtering will be done, one-sided limits, eg. \verb{(NA, 10)} or \verb{(10, NA)}
implies to filter \code{<=} or \code{>=}, respectively.}
}
\value{
function
}
\description{
Generate a scan filtering function
}
\examples{
\dontrun{
  # filter retention time and y-frequency coefficient
  sc_filter = generate_scan_outlier_filter(rtime = c(0, 450), y.freq = c(2.9e7, NA))

  # one sided filters
  sc_filter = generate_scan_outlier_filter(rtime = c(NA, 450), y.freq = c(2.9e7, NA))

  # no filtering
  sc_filter = generate_scan_outlier_filter()
}
}
