% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{get_area_peak_rejslope}
\alias{get_area_peak_rejslope}
\title{get area peak}
\usage{
get_area_peak_rejslope(possible_peak, min_points)
}
\arguments{
\item{possible_peak}{the peak data}

\item{min_points}{how many points needed}
}
\value{
numeric
}
\description{
get area peak
}
