% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{get_cauchy_peak_info}
\alias{get_cauchy_peak_info}
\title{generate peak info using a cauchy fit}
\usage{
get_cauchy_peak_info(possible_peak, w = NULL)
}
\arguments{
\item{possible_peak}{data.frame of mz, intensity}

\item{w}{the weightes to use for the points}
}
\value{
data.frame
}
\description{
generate peak info using a cauchy fit
}

