% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMS.R
\name{get_ms_info}
\alias{get_ms_info}
\title{get MS info}
\usage{
get_ms_info(raw_data, include_msn = FALSE, include_precursor = FALSE)
}
\arguments{
\item{raw_data}{the xcms raw data object}

\item{include_msn}{should information from MSn scans be included?}

\item{include_precursor}{should the precursor scans be included?}
}
\value{
data.frame with scan, time, acquisition, tic, ms_level and ms_type
}
\description{
get MS info
}
