% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{get_mzml_header}
\alias{get_mzml_header}
\title{extract mzML header}
\usage{
get_mzml_header(mzml_file)
}
\arguments{
\item{mzml_file}{the mzML file to get the header from}
}
\description{
extract mzML header
}
