% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{get_mzml_metadata}
\alias{get_mzml_metadata}
\title{get mzML metadata}
\usage{
get_mzml_metadata(mzml_file)
}
\arguments{
\item{mzml_file}{the mzML file to get metadata from}
}
\description{
get mzML metadata
}
