% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{get_peak_info}
\alias{get_peak_info}
\title{get peak info}
\usage{
get_peak_info(peak_data, peak_method = "lm_weighted", min_points = 4)
}
\arguments{
\item{peak_data}{data.frame with \code{mz}, \code{intensity}, and log intensity \code{log_int}}

\item{peak_method}{which method to use to get information about the peak (see Details)}

\item{min_points}{how many points have to still be in a peak to return something useful?}
}
\value{
data.frame
}
\description{
Given a set of points that make up a peak, use the provided method to find
the center, intensity and area of the peak.
}
\details{
This function can choose from several methods to get information
about the peak. Currently available methods include:

\describe{
\item{lm_weighted}{quadratic linear fit to log-intensities, with weights for
each point relative to the maximum intensity point (current default)}
\item{lm_unweighted}{quadratic linear fit to log-intensities}
}
}
