% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{get_raw_ms_metadata}
\alias{get_raw_ms_metadata}
\title{get raw ms metadata}
\usage{
get_raw_ms_metadata(in_file)
}
\arguments{
\item{in_file}{the file to use}
}
\value{
list
}
\description{
figures out which metadata function to run, and returns back the metadata
generated by it.
}

