% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{get_rsq_peak}
\alias{get_rsq_peak}
\title{get rsq peak}
\usage{
get_rsq_peak(possible_peak, min_rsq, min_points)
}
\arguments{
\item{possible_peak}{the peak data}

\item{min_rsq}{minimum rsq}

\item{min_points}{min number of points}
}
\value{
numeric
}
\description{
get rsq peak
}

