% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRaw.R
\name{get_scan_info}
\alias{get_scan_info}
\title{get MS info}
\usage{
get_scan_info(raw_data)
}
\arguments{
\item{raw_data}{the MSnbase raw data object}

\item{include_msn}{should information from MSn scans be included?}

\item{include_precursor}{should the precursor scans be included?}
}
\value{
data.frame with scan, time, acquisition, tic, ms_level and ms_type
}
\description{
get MS info
}
