% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{get_scan_nozeros}
\alias{get_scan_nozeros}
\title{filter raw scan data}
\usage{
get_scan_nozeros(rawdata, cutoff = 0.00225)
}
\arguments{
\item{rawdata}{data.frame of mz and intensity}

\item{cutoff}{the maximum difference value to consider}
}
\description{
filter the raw scan data so we can use it for resolution
}
