% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{import_mzML}
\alias{import_mzML}
\title{import mzML data}
\usage{
import_mzML(mz_file, profstep = 0, includeMSn = TRUE)
}
\arguments{
\item{mz_file}{the mzML file to import}

\item{profstep}{the profile step to use, should be 0 to make our lives easier}

\item{includeMSn}{whether to include MSn data, should be TRUE to get \emph{all data}.}

\item{...}{other xcmsRaw parameters}
}
\value{
xcmsRaw
}
\description{
function to import mzML data in a way that provides what we need to work
with it. \code{mz_file} should be the \emph{full path} to the data.
}

