% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRaw.R
\name{import_sc_raw}
\alias{import_sc_raw}
\title{import raw mass spec data}
\usage{
import_sc_raw(raw_data, ms_level = 1)
}
\arguments{
\item{raw_data}{the raw mass spec file to import}

\item{ms_level}{which MS-level data to import}
}
\value{
xcmsRaw
}
\description{
function to import raw mass spec data in a way that provides what we need to work
with it. \code{raw_data} should be the \emph{full path} to the data.
}
