% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{import_xcalibur}
\alias{import_xcalibur}
\title{import Xcalibur data}
\usage{
import_xcalibur(xcal_file, sheet = 1)
}
\arguments{
\item{xcal_file}{the Xcalibur file to import}

\item{sheet}{which sheet to read}
}
\value{
tbl_df
}
\description{
function to import Xcalibur data with the peak locations
}
