% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{initialize_metadata_from_mzml}
\alias{initialize_metadata_from_mzml}
\title{initialize metadata from mzML}
\usage{
initialize_metadata_from_mzml(zip_dir, mzml_file)
}
\arguments{
\item{zip_dir}{the directory containing unzipped data}

\item{mzml_file}{the mzML file to extract metadata from}
}
\description{
initialize metadata from mzML
}
