% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{initialize_zip_metadata}
\alias{initialize_zip_metadata}
\title{initialize metadata}
\usage{
initialize_zip_metadata(zip_dir)
}
\arguments{
\item{zip_dir}{the temp directory that represents the final zip}
}
\description{
initialize metadata
}
