% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{integrate_model}
\alias{integrate_model}
\title{integrate model}
\usage{
integrate_model(model_mz, model_coeff, n_point = 100,
  log_transform = "log")
}
\arguments{
\item{model_mz}{the mz values for the model peak}

\item{model_coeff}{the model of the peak}

\item{n_point}{how many points to use for integration}

\item{log_transform}{what kind of transform was applied}
}
\value{
numeric
}
\description{
provides the area integration for the peak that fits the parabolic model
}
