% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{integrate_sides}
\alias{integrate_sides}
\title{integrate sides}
\usage{
integrate_sides(peak_mz, peak_int, full_peak_loc, model_peak_loc)
}
\arguments{
\item{peak_mz}{the mz in the peak}

\item{peak_int}{the intensity in the peak}

\item{full_peak_loc}{what defines all of the peak}

\item{model_peak_loc}{what defined the peak fitting the parabolic model}
}
\value{
numeric
}
\description{
provides ability to calculate the area on the sides of a peak that are not
caught by the parabolic model assuming a triangle to each side of the parabola
}

