% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{json_2_peak_list}
\alias{json_2_peak_list}
\title{PeakList from json}
\usage{
json_2_peak_list(json_string, in_var = "Peaks")
}
\arguments{
\item{json_string}{the json to convert}

\item{in_var}{the top level variable containing the "Peaks"}
}
\value{
tbl_df
}
\description{
takes json representing a PeakList object, and generates the data.frame version
}

