% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{json_mzML_2_df}
\alias{json_mzML_2_df}
\title{json mzML to data.frame}
\usage{
json_mzML_2_df(in_file)
}
\arguments{
\item{in_file}{the file to read from}
}
\value{
data.frame
}
\description{
Given a json file or list of lists, return a data.frame with the most important
bits of the data.
}
