% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{linear_fit}
\alias{linear_fit}
\title{linear fit}
\usage{
linear_fit(x, y)
}
\arguments{
\item{x}{the x-values, the independent value}

\item{y}{the y-values, the dependent value}
}
\value{
vector
}
\description{
calculates system of equations for a linear fit
}

