% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{lists_2_json}
\alias{lists_2_json}
\title{lists_2_json}
\usage{
lists_2_json(lists_to_save, zip_file = NULL, digits = 8,
  temp_dir = tempfile(pattern = "json"))
}
\arguments{
\item{lists_to_save}{the set of lists to create the json from}

\item{zip_file}{should the JSON files be zipped into a zip file? Provide the zip file name}

\item{digits}{how many digits to use for the JSON representation}

\item{temp_dir}{temp directory to write the JSON files to}
}
\value{
character
}
\description{
lists_2_json
}
