% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{load_metadata}
\alias{load_metadata}
\title{load metadata}
\usage{
load_metadata(zip_dir, metadata_file)
}
\arguments{
\item{zip_dir}{the directory of the unzipped data}

\item{metadata_file}{the metadata file}
}
\value{
list
}
\description{
given a zip and a metadata file, load it and return it
}
