% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{loess_to_df}
\alias{loess_to_df}
\title{create loess data.frame}
\usage{
loess_to_df(loess_model)
}
\arguments{
\item{loess_model}{the model object generated by loess}
}
\value{
data.frame
}
\description{
Given a loess model, creates a data.frame suitable for plotting via ggplot2
}
