% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{log_peaks}
\alias{log_peaks}
\title{generate log intensity}
\usage{
log_peaks(mz_data, intensity = "intensity")
}
\arguments{
\item{mz_data}{a data.frame of mz's and intensities}

\item{intensity}{which column name has the intensities}
}
\value{
data.frame
}
\description{
given some m/z intensity data, log-transform with some small constant added
}
