% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{MasterPeakList}
\alias{MasterPeakList}
\title{Storing a master list of peaks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MasterPeakList
}
\description{
Stores a master list of peaks, and which peaks from which scan match it
}
\keyword{datasets}

