% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{MasterSampleList}
\alias{MasterSampleList}
\title{MasterSampleList}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MasterSampleList
}
\description{
Holds the correspondence across samples
}
\keyword{datasets}
