% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\name{meta_export_json}
\alias{meta_export_json}
\title{export metadata to json}
\usage{
meta_export_json(meta_list)
}
\arguments{
\item{meta_list}{a list of metadata}
}
\description{
export the list metadata to a json string
}

