% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{model_peak_center_intensity}
\alias{model_peak_center_intensity}
\title{model peak center}
\usage{
model_peak_center_intensity(x, coefficients)
}
\arguments{
\item{x}{the x-values to use (non-centered)}

\item{coefficients}{the model coefficients generated from centered model}
}
\value{
numeric
}
\description{
use the derivative of the parabolic equation to find the peak center,
and then put the center into the equation to find the intensity at that
point.
}
\details{
The coefficients are generated using the linear model:
\deqn{y = a + bx + cx^2}.

The derivative of this is:
\deqn{y = b + 2cx}

The peak of a parabola is defined where y is zero for the derivative.

\deqn{x = -b / 2c}

We can use this to derive where the center of the peak is, and then put the
center value back into the equation to get the intensity.
}
