% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{MultiSamplePeakList}
\alias{MultiSamplePeakList}
\title{MultiSamplePeakList}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiSamplePeakList
}
\description{
PeakList objects for multiple samples
}
\keyword{datasets}
