% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{MultiScans}
\alias{MultiScans}
\title{multiple ms scans}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiScans
}
\description{
stores the results of peak picking on multiple MS scans
}
\keyword{datasets}

