% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{MultiScansPeakList}
\alias{MultiScansPeakList}
\title{multiple scans peak lists}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiScansPeakList
}
\description{
holds a list of peak lists where each list entry is from a different scan
}
\keyword{datasets}

