% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{mz_scans_to_frequency}
\alias{mz_scans_to_frequency}
\title{convert mz to frequency across scans}
\usage{
mz_scans_to_frequency(mz_scan_df)
}
\arguments{
\item{mz_scan_df}{a data.frame with at least \code{mz} and \code{scan} columns}

\item{...}{other parameters for \code{convert_mz_frequency}}
}
\value{
data.frame
}
\description{
Given a multi-scan data.frame of m/z, generate frequency values for the data.
}
\seealso{
convert_mz_frequency
}
