% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{mzml_to_zip}
\alias{mzml_to_zip}
\title{create initial zip from mzML}
\usage{
mzml_to_zip(mzml_file, out_file)
}
\arguments{
\item{mzml_file}{the mzML file to zip up}

\item{out_file}{the directory to save the zip file}
}
\description{
given an mzML file, create the initial zip file containing the
zipped \emph{mzML}, \emph{metadata.json}, and \emph{mzml_metadata.json}.
This zip file is what will be operated on by anything that accesses files,
so that our interface is consistent.
}
