% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{noise_detector}
\alias{noise_detector}
\title{noise_detector}
\usage{
noise_detector(peaklist, intensity_measure = "Height", transform = log10)
}
\arguments{
\item{peaklist}{data.frame of peaks}

\item{intensity_measure}{which column of the data.frame is the intensity to use}

\item{transform}{which transform to apply to the data}
}
\value{
list
}
\description{
Given a peak_list as a data.frame, classify the peaks as either noise or
signal. See Details for more information.
}
\details{
This noise detections works by assuming that noise and signal
 come from two distinct distributions, and that the noise distribution
 has a lower valued distribution than the signal. So to run, first it calculates
 a density or smoothed histogram on the intensity values. Second, the first
 peak becomes the mean value of the noise. All peak intensities below this value
 are used to calculate the standard deviation of the noise, and then the mean
 + 3 sd is used to define the upper bound of the noise intensity.

 Very importantly, you should graph a histogram of your intensities to verify
 that a transform is necessary. This is written for data with a Poisson type
 variance structure, so "log10" is appropriate. Check that your data meets these
 assumptions before using this transform.

 Finally, this will return a list with two components, the original peak_list
 data.frame with a logical column "not_noise" appended, and summary information
 about the noise values themselves.
}
