% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{noise_sorted_peaklist}
\alias{noise_sorted_peaklist}
\title{noise from peaklist}
\usage{
noise_sorted_peaklist(peaklist, intensity_measure = "Height",
  sd_mean_ratio = 1.2, noise_multiplier = 2)
}
\arguments{
\item{peaklist}{the data.frame with at least "Height" or "Area"}

\item{intensity_measure}{which value of \emph{intensity} should be used? Default is "Height"}

\item{sd_mean_ratio}{the ratio of standard deviation to mean to use as a cutoff}

\item{noise_multiplier}{how high above the noise should the cutoff be}
}
\value{
list
}
\description{
Given a peak list data.frame object, determines a noise cutoff and classifies
peaks as to whether they are "not_noise".
}
\details{
This calculation is based on the premise that a distribution of
 only noise peaks should have a standard deviation to mean ratio of about 1.
 Therefore, but sorting the list of peak intensities, and incrementally adding
 peaks, when the sd to mean ratio becomes larger than the cutoff, we say
 that we are out of the noise region of the data.

 The multiplier determines how much higher than the median of the noise do
 we want to consider a peak as \emph{real}, i.e. definitely not noise.

 A list is returned, with the original data.frame with a new column "not_noise",
 and the actual noise cutoff that was used to determine whether peaks are noise.
}
