% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{normalize_mspl}
\alias{normalize_mspl}
\title{normalize MultiScanPeakList}
\usage{
normalize_mspl(normalization_factors, mspl)
}
\arguments{
\item{normalization_factors}{the factors used for normalization}

\item{mspl}{the \code{MultiScanPeakList} object that needs normalization}
}
\value{
MultiScanPeakList
}
\description{
Given a set of normalization factors and a \code{MultiScanPeakList}, normalize
the peaks in each scan of the \code{MultiScanPeakList}
}
