% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{normalize_scans}
\alias{normalize_scans}
\title{normalize scans}
\usage{
normalize_scans(mpl, intensity_measure = "Height",
  summary_function = mean)
}
\arguments{
\item{mpl}{the MasterPeakList object}

\item{intensity_measure}{which measure of peak intensity to use?}

\item{summary_function}{which function to use to summarize the differences}
}
\description{
Given a \code{MasterPeakList} object that has the peaks across scans corresponded,
normalize the scans against each other.
}
\details{
To do the normalization, we find the set of peaks that have a large
number of corresponding peaks, and using these peaks, for each peak, for each
scan get the log-ratio of the peak in that scan against all the other scans
}
