% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{parabolic_fit}
\alias{parabolic_fit}
\title{parabolic fit}
\usage{
parabolic_fit(x, y, w = NULL)
}
\arguments{
\item{x}{the x-values, independent}

\item{y}{the y-values, dependent}

\item{w}{weights}
}
\value{
list
}
\description{
calculates the coefficients of a parabolic fit (y = x + x^2) of x to y
}
