% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{peak_finder}
\alias{peak_finder}
\title{peak finding and reporting}
\usage{
peak_finder(raw_data, method = "lm_weighted",
  noise_function = noise_sorted_peaklist)
}
\value{
list
}
\description{
Given a RawMS object, actually does peak finding at the scan level and then
peak correspondence.
}
