% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{peak_info2}
\alias{peak_info2}
\title{peak info2}
\usage{
peak_info2(possible_peak, min_points = 5, min_area = 0.1)
}
\arguments{
\item{possible_peak}{data.frame of mz, intensity and log intensity (log_int)}

\item{min_points}{the minimum number of points in a peak}

\item{min_area}{the minimum area for a peak}
}
\value{
numeric
}
\description{
given a peak found, try to fit a parabolic model and return the center,
intensity, and area using basic and model based if possible.
}

