% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\name{peak_list_2_json}
\alias{peak_list_2_json}
\title{PeakList to json}
\usage{
peak_list_2_json(peak_list)
}
\arguments{
\item{peak_list}{a data.frame or tbl_df to convert}
}
\value{
json_string
}
\description{
takes a PeakList object, and generates a json version
}
