% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeMS.R
\docType{data}
\name{PeakFinder}
\alias{PeakFinder}
\title{R6 peak_finder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PeakFinder
}
\description{
an R6 based peak_finder class that controls execution in a more fine grained
manner, and allows the possibility of easily checking each step and saving
intermediate outputs when and if desired.
}
\keyword{datasets}
