% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{PeakList}
\alias{PeakList}
\title{peaklist}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PeakList
}
\description{
store a peaklist, which is a data.frame of ObservedMZ, Height and Area,
and which scan it is from, and the noise level.
}
\keyword{datasets}

