% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{PeakMS}
\alias{PeakMS}
\title{Storing a peak and associated statistics}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PeakMS
}
\arguments{
\item{peak_data}{the actual mz, intensity, and log-intensity of the peak}
}
\value{
a reference class with summarized information about the peak
}
\description{
This reference class is a storage container for a mass-spec peak, and knows
how to pass the data to the various summary functions for finding the peak
center, intensity, area, etc.
}
\keyword{internal}

