% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPicking.R
\docType{data}
\name{PeakPickingAnalysis}
\alias{PeakPickingAnalysis}
\title{peak picking analysis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PeakPickingAnalysis
}
\description{
Reference class to hold the results of a peak picking analysis
}
\keyword{datasets}

