% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing_peaks.R
\name{plot_peaks}
\alias{plot_peaks}
\title{plot peaks}
\usage{
plot_peaks(peak_lists, mean_int, range_plusminus = 1e-06)
}
\arguments{
\item{mean_int}{the mean intensities to also plot}

\item{peak_list}{the tbl_df of peaks}
}
\value{
ggplot
}
\description{
generates a plot of peaks so we can compare them
}
