% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{plot_raw_peaks}
\alias{plot_raw_peaks}
\title{plot raw peaks}
\usage{
plot_raw_peaks(sc_raw, scan_range = NULL, mz_range = NULL, transform = NULL)
}
\arguments{
\item{sc_raw}{either a RawMS object, or data.frame of mz / intensity}

\item{scan_range}{which scans to use}

\item{mz_range}{the range of M/z's to consider}

\item{transform}{apply a transform to the data}
}
\value{
ggplot2 object
}
\description{
Given the sc_raw object, generate a plot of the peaks that are generated from
averaging the scans in the currently set scan-range.
}
