% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMS.R
\name{plot_tic}
\alias{plot_tic}
\title{plot tic}
\usage{
plot_tic(raw_data, color_ms = TRUE, log_transform = TRUE)
}
\arguments{
\item{raw_data}{an MSnbase raw object}

\item{color_ms}{should scans be colored by their \emph{ms} level and type?}
}
\value{
ggplot
}
\description{
function to plot the total intensity chromatogram of the data, with information
about which scans are which
}
