% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{pracma_findpeaks}
\alias{pracma_findpeaks}
\title{return peaks}
\usage{
pracma_findpeaks(avg_spectra, ...)
}
\arguments{
\item{avg_spectra}{the avg spectra data}

\item{...}{other findpeaks parameters}
}
\value{
tbl_df
}
\description{
returns peaks found by \code{pracma::findpeaks}
}

