% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{predicted_frequency_r2}
\alias{predicted_frequency_r2}
\title{calculate r2}
\usage{
predicted_frequency_r2(mz_frequency_df)
}
\arguments{
\item{mz_frequency_df}{the data.frame with predicted frequencies}
}
\description{
calculate r2
}
