% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{r_squared}
\alias{r_squared}
\title{check fit}
\usage{
r_squared(x, y)
}
\arguments{
\item{x}{numeric x values}

\item{y}{numeric y values}
}
\value{
numeric
}
\description{
given a series of x's and y's, how well do they fit a parabolic model?
}

