% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{raw_metadata_mzml}
\alias{raw_metadata_mzml}
\title{get raw metadata}
\usage{
raw_metadata_mzml(mzml_files, raw_file_loc, recursive = TRUE)
}
\arguments{
\item{mzml_files}{the paths to the mzml files}

\item{raw_file_loc}{the directory holding raw files and json metadata files}
}
\description{
When raw files are copied, we also generated metadata about their original locations
and new locations, and some other useful info. We would like to capture it, and
keep it along with the metadata from the mzml file. So, given a list of mzml
files, and a location for the raw files, this function creates metadata json
files for the mzml files.
}
