% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_ms.R
\name{raw_ms}
\alias{raw_ms}
\title{Storing the raw mass spec data and operating on it}
\usage{
raw_ms(in_file, out_dir = dirname(in_file))
}
\arguments{
\item{in_file}{the zip or mzML file to use}

\item{out_dir}{which directory a new zip file should be in}
}
\value{
A reference class with methods \code{set_scans()}, \code{plot_tic()},
 \code{find_peaks()}.
}
\description{
This reference class represents a storage container for the raw mass-spec
data from a zipped mzML file and associated metadata. It provides methods
for plotting the \emph{total-ion chromatogram}, setting which scans will
be used for peak picking, as well as setting which function will be used
to do peak picking.
}
\details{
The \code{find_peaks()} method is not set by default, it must be set
 by the user after instantiation. The \code{find_peaks()} function definition
 should take the \code{raw_data}, and either \code{scan_range} or \code{rt_range}.
 It should also store the function name that is actually used for finding the peaks,
 and the variables, and then generate a \code{peaks} object. See the examples
 below for how to generate these functions.
}
\examples{

\dontrun{
# setting different files
rms <- raw_ms("mzmlFile.mzML")
rms <- raw_ms("mzmlFile.mzML", "someDirectoryForZip")

rms <- raw_ms("mzmlZipFile.zip")

# Plot the total-ion-chromatogram
rms$plot_tic()

rms$set_scans(scan_range = c(1, 2, 3))
rms$set_scans(rt_range = c(0, 50.1)) # uses seconds

# set the find_peaks method
rms$find_peaks <- function(){

}

}

}
\keyword{internal}

