% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{raw_peak_intensity}
\alias{raw_peak_intensity}
\title{raw peaks intensities}
\usage{
raw_peak_intensity(raw_ms, scanrange = raw_ms$scan_range)
}
\arguments{
\item{raw_ms}{an RawMS object}

\item{scanrange}{the range of scans to use, default is derived from raw_ms}
}
\value{
data.frame of mz and intensity
}
\description{
generate raw peaks from just averaging scans via xcms
}
