% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{raw_peaks}
\alias{raw_peaks}
\title{raw peaks}
\usage{
raw_peaks(sc_raw, scanrange = sc_raw$scan_range)
}
\arguments{
\item{sc_raw}{an RawMS object}

\item{scanrange}{the range of scans to use, default is derived from sc_raw}
}
\value{
numeric
}
\description{
generate raw peaks from just averaging scans via xcms
}
