% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawMS.R
\docType{data}
\name{RawMS}
\alias{RawMS}
\title{Storing the raw mass spec data and operating on it}
\format{
An object of class \code{R6ClassGenerator} of length 25.
}
\usage{
RawMS
}
\arguments{
\item{raw_file}{the zip or mzML file to use}

\item{metadata_file}{the metadata file to use}
}
\value{
A reference class with methods \code{set_scans()}, \code{plot_tic()},
\code{find_peaks()}.
}
\description{
This reference class represents a storage container for the raw mass-spec
data from a zipped mzML file and associated metadata. It provides methods
for plotting the \emph{total-ion chromatogram}, and setting which scans will
be used for peak picking
}
\details{
The \code{find_peaks()} method is not set by default, it must be set
by the user after instantiation. The \code{find_peaks()} function definition
should take the \code{raw_data}, and either \code{scan_range} or \code{rt_range}.
It should also store the function name that is actually used for finding the peaks,
and the variables, and then generate a \code{peaks} object. See the examples
below for how to generate these functions.
}
\examples{

\dontrun{
# setting different files
rms = raw_ms("mzmlFile.mzML")
rms = raw_ms("mzmlFile.mzML", "someDirectoryForZip")

rms = raw_ms("mzmlZipFile.zip")

# Plot the total-ion-chromatogram
rms$plot_tic()

rms$set_scans(scan_range = c(1, 2, 3))
rms$set_scans(rt_range = c(0, 50.1)) # uses seconds

# set the find_peaks method
rms$find_peaks = function(){

}

}

}
\keyword{internal}
