% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{recalculate_offsets}
\alias{recalculate_offsets}
\title{recalculate offsets}
\usage{
recalculate_offsets(in_zip, offset = 2, out_file = in_zip)
}
\arguments{
\item{in_zip}{the zip file to work with}

\item{offset}{the offset to use}

\item{out_file}{the file to write too (optional)}
}
\description{
Given a previously generated zip file of characterized peaks, now we've realized
that the offsets on each peak should be somehow different. This function takes
a zip file, adjusts the offsets, and writes the file back out.
}
