% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_peaks.R
\name{refine_closest_list}
\alias{refine_closest_list}
\title{refine closest}
\usage{
refine_closest_list(closest_list)
}
\arguments{
\item{closest_list}{input data.frame}
}
\value{
list
}
\description{
after finding closest peaks, goes through and looks for cases where there are
multiple matches, and selects the smallest difference to keep.
}

