% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_mzml_list}
\alias{run_mzml_list}
\title{run list of mzML files}
\usage{
run_mzml_list(
  mzml_files,
  json_files = NULL,
  progress = TRUE,
  save_loc = ".",
  ...
)
}
\arguments{
\item{mzml_files}{the list of mzML files to use}

\item{json_files}{the list of corresponding json meta-data files}

\item{save_loc}{where should the file files be saved}

\item{...}{other parameters for \code{CharacterizeMS}}
}
\value{
list
}
\description{
Given a data.frame or character vector of files to run characterization
on, processes them in sequence, to a particular saved location.
}
