% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{sample_run_time}
\alias{sample_run_time}
\title{determine sample run time}
\usage{
sample_run_time(zip, units = "m")
}
\arguments{
\item{zip}{the zip object you want to use}

\item{units}{what units should the run time be in? (s, m, h)}
}
\value{
data.frame with sample, start and end time
}
\description{
determine sample run time
}
