% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{sample_ti.MasterPeakList}
\alias{sample_ti.MasterPeakList}
\title{get sample ti}
\usage{
\method{sample_ti}{MasterPeakList}(x, min_scan = 2)
}
\arguments{
\item{x}{the object}

\item{min_scan}{how many scans should a peak be in to contribute?}
}
\value{
data.frame
}
\description{
Calculate the sample total intensities
}
