% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{save}
\alias{save}
\title{SCZip - save}
\arguments{
\item{out_file}{the file to save to}
}
\description{
SCZip - save
}
\details{
\code{out_file}, if it is \code{NULL}, will be taken from when the
object was generated, and by default will be set to the same as the \code{in_file}.
If not \code{NULL}, then it is checked that the \code{id} is part of the
\code{out_file}, and if not, the \code{id} is added to the actual file name.
}
\examples{
\dontrun{
 new_ms = sc_zip("in_file")
 new_ms$save()
 new_ms$save("out_file")
}
}
