% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{scan_mzML}
\alias{scan_mzML}
\title{import mzML by scans}
\usage{
scan_mzML(mz_file, scan_indices = NULL, scan_times = NULL,
  mz_range = numeric())
}
\arguments{
\item{mz_file}{the mzML file}

\item{scan_indices}{which scans to import by index}

\item{scan_times}{which scans to import by time (in seconds)}

\item{mz_range}{a range of m/z's to return the scans by}
}
\value{
list
}
\description{
given an mzML file, import all or a range of scans
}
\details{
Returns a list of \code{tbl_df}'s, with \code{mz} and \code{intensity}.
}
\examples{
\dontrun{
  mz_file <- "example_file.mzML"
  scan_data <- scan_mzML(mz_file) # imports all scans
  scan_data <- scan_mzML(mz_file, scan_indices = seq(1, 5)) # scans 1-5
  scan_data <- scan_mzML(mz_file, scan_times = c(24, 80)) # scans in this time range
}
}
