% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{scan_time_filter}
\alias{scan_time_filter}
\title{filter scans using time}
\usage{
scan_time_filter(raw_ms, min_time_difference = 4)
}
\arguments{
\item{raw_ms}{a RawMS object}

\item{min_time_difference}{a minimum cutoff for the time difference in minutes}
}
\value{
RawMS
}
\description{
Implements a simple scan filter that considers that time it took to acquire
the scan. Useful when there is a mixture of MS1 scans, some acquired at
different resolutions or different numbers of microscans.
}
