% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScanCentricPeakCharacterization.R
\docType{package}
\name{ScanCentricPeakCharacterization}
\alias{ScanCentricPeakCharacterization}
\title{ScanCentricPeakCharacterization: A package for direct injection FT-MS data processing.}
\description{
The ScanCentricPeakCharacterization package provides several classes and functions
for working with direct injection, high-resolution mass spectrometry data.
}
