% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\docType{data}
\name{ScanMS}
\alias{ScanMS}
\title{Storing all of the peaks associated with a scan}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScanMS
}
\arguments{
\item{scan_data}{a data.frame of mz and intensity}

\item{min_points}{how many points to consider a "peak"}

\item{n_peak}{how many peaks are allowed to be found}

\item{flat_cut}{relative intensity to decide a peak is "flat"}
}
\value{
reference class with a list of peaks
}
\description{
Reference class to hold the results of peak finding of an entire "scan"
within a mass-spectrum.
}
\keyword{datasets}
