% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{scans_to_json}
\alias{scans_to_json}
\title{scans to json}
\usage{
scans_to_json(peak_finder, exclude_noise = TRUE, file_output = NULL)
}
\arguments{
\item{peak_finder}{the peak_finder object that has everything}

\item{exclude_noise}{should noise peaks be excluded?}

\item{file_output}{where should the json go?}
}
\value{
character
}
\description{
write the scan level peaks to a JSON format
}
