% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCharacterizePeaks.R
\name{SCCharacterizePeaks}
\alias{SCCharacterizePeaks}
\title{analyze scan-centric mass-spec data}
\description{
This class allows you to analyze mass spec data, and controls the execution
of reading in the mass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_file}{\code{SCCharacterizePeaks$load_file()}}
\item \href{#method-prepare_raw_data}{\code{SCCharacterizePeaks$prepare_raw_data()}}
\item \href{#method-check_frequency_model}{\code{SCCharacterizePeaks$check_frequency_model()}}
\item \href{#method-get_frequency_data}{\code{SCCharacterizePeaks$get_frequency_data()}}
\item \href{#method-scan_info}{\code{SCCharacterizePeaks$scan_info()}}
\item \href{#method-find_peaks}{\code{SCCharacterizePeaks$find_peaks()}}
\item \href{#method-summarize}{\code{SCCharacterizePeaks$summarize()}}
\item \href{#method-save_peaks}{\code{SCCharacterizePeaks$save_peaks()}}
\item \href{#method-write_zip}{\code{SCCharacterizePeaks$write_zip()}}
\item \href{#method-set_peak_finder}{\code{SCCharacterizePeaks$set_peak_finder()}}
\item \href{#method-run_all}{\code{SCCharacterizePeaks$run_all()}}
\item \href{#method-prep_data}{\code{SCCharacterizePeaks$prep_data()}}
\item \href{#method-add_regions}{\code{SCCharacterizePeaks$add_regions()}}
\item \href{#method-run_splitting}{\code{SCCharacterizePeaks$run_splitting()}}
\item \href{#method-new}{\code{SCCharacterizePeaks$new()}}
\item \href{#method-clone}{\code{SCCharacterizePeaks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-load_file}{}}}
\subsection{Method \code{load_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$load_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_raw_data"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_raw_data}{}}}
\subsection{Method \code{prepare_raw_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$prepare_raw_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_frequency_model"></a>}}
\if{latex}{\out{\hypertarget{method-check_frequency_model}{}}}
\subsection{Method \code{check_frequency_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$check_frequency_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_frequency_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_frequency_data}{}}}
\subsection{Method \code{get_frequency_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$get_frequency_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_info"></a>}}
\if{latex}{\out{\hypertarget{method-scan_info}{}}}
\subsection{Method \code{scan_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$scan_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-find_peaks}{}}}
\subsection{Method \code{find_peaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$find_peaks(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-save_peaks}{}}}
\subsection{Method \code{save_peaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$save_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_zip"></a>}}
\if{latex}{\out{\hypertarget{method-write_zip}{}}}
\subsection{Method \code{write_zip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$write_zip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_peak_finder"></a>}}
\if{latex}{\out{\hypertarget{method-set_peak_finder}{}}}
\subsection{Method \code{set_peak_finder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$set_peak_finder(in_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$run_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prep_data"></a>}}
\if{latex}{\out{\hypertarget{method-prep_data}{}}}
\subsection{Method \code{prep_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$prep_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_regions"></a>}}
\if{latex}{\out{\hypertarget{method-add_regions}{}}}
\subsection{Method \code{add_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$add_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_splitting"></a>}}
\if{latex}{\out{\hypertarget{method-run_splitting}{}}}
\subsection{Method \code{run_splitting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$run_splitting()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$new(
  in_file,
  metadata_file = NULL,
  out_file = NULL,
  temp_loc = NULL,
  frequency_fit_description = NULL,
  mz_fit_description = NULL,
  filter_remove_outlier_scans = NULL,
  choose_single_frequency_model = NULL,
  sc_peak_region_finder = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
