% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{SCPeakRegionFinder}
\alias{SCPeakRegionFinder}
\title{R6 Peak Region Finder}
\description{
R6 Peak Region Finder

R6 Peak Region Finder
}
\details{
Think of it like managing all the stuff that needs to happen to find the peaks
in the regions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{run_time}}{how long did the process take}

\item{\code{start_time}}{when did we start}

\item{\code{stop_time}}{when did we start}

\item{\code{peak_regions}}{SCPeakRegions object}

\item{\code{sliding_region_size}}{how big are the sliding regions in data points}

\item{\code{sliding_region_delta}}{how much space between sliding region starts}

\item{\code{quantile_multiplier}}{how much to multiply quantile based cutoff by}

\item{\code{n_point_region}}{how many points are there in the big tiled regions for quantile based cutoff}

\item{\code{tiled_region_size}}{how wide are the tiled regions in data points}

\item{\code{tiled_region_delta}}{how far in between each tiled region}

\item{\code{region_percentile}}{??}

\item{\code{peak_method}}{what method to extract peak center, height, area, etc}

\item{\code{min_points}}{how many points wide does a peak have to be to get characterized}

\item{\code{sample_id}}{what sample are we processing}

\item{\code{n_zero_tiles}}{how many zero count tiled regions split up a region into multiple peaks?}

\item{\code{zero_normalization}}{do we want to pretend to do normalization}

\item{\code{calculate_peak_area}}{should peak area be calculated as well?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_regions}{\code{SCPeakRegionFinder$add_regions()}}
\item \href{#method-reduce_sliding_regions}{\code{SCPeakRegionFinder$reduce_sliding_regions()}}
\item \href{#method-split_peak_regions}{\code{SCPeakRegionFinder$split_peak_regions()}}
\item \href{#method-remove_double_peaks_in_scans}{\code{SCPeakRegionFinder$remove_double_peaks_in_scans()}}
\item \href{#method-normalize_data}{\code{SCPeakRegionFinder$normalize_data()}}
\item \href{#method-find_peaks_in_regions}{\code{SCPeakRegionFinder$find_peaks_in_regions()}}
\item \href{#method-model_mzsd}{\code{SCPeakRegionFinder$model_mzsd()}}
\item \href{#method-model_heightsd}{\code{SCPeakRegionFinder$model_heightsd()}}
\item \href{#method-indicate_high_frequency_sd}{\code{SCPeakRegionFinder$indicate_high_frequency_sd()}}
\item \href{#method-add_data}{\code{SCPeakRegionFinder$add_data()}}
\item \href{#method-summarize_peaks}{\code{SCPeakRegionFinder$summarize_peaks()}}
\item \href{#method-add_offset}{\code{SCPeakRegionFinder$add_offset()}}
\item \href{#method-sort_ascending_mz}{\code{SCPeakRegionFinder$sort_ascending_mz()}}
\item \href{#method-characterize_peaks}{\code{SCPeakRegionFinder$characterize_peaks()}}
\item \href{#method-summarize}{\code{SCPeakRegionFinder$summarize()}}
\item \href{#method-peak_meta}{\code{SCPeakRegionFinder$peak_meta()}}
\item \href{#method-new}{\code{SCPeakRegionFinder$new()}}
\item \href{#method-clone}{\code{SCPeakRegionFinder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_regions"></a>}}
\if{latex}{\out{\hypertarget{method-add_regions}{}}}
\subsection{Method \code{add_regions()}}{
Add the sliding and tiled regions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$add_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce_sliding_regions"></a>}}
\if{latex}{\out{\hypertarget{method-reduce_sliding_regions}{}}}
\subsection{Method \code{reduce_sliding_regions()}}{
Find the regions most likely to contain real signal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$reduce_sliding_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split_peak_regions"></a>}}
\if{latex}{\out{\hypertarget{method-split_peak_regions}{}}}
\subsection{Method \code{split_peak_regions()}}{
Split up signal regions by peaks found
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$split_peak_regions(use_regions = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_regions}}{an index of the regions we want to split up}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_double_peaks_in_scans"></a>}}
\if{latex}{\out{\hypertarget{method-remove_double_peaks_in_scans}{}}}
\subsection{Method \code{remove_double_peaks_in_scans()}}{
Check for the presence of two peaks with the same scan number in each
region and remove them. Any regions with zero peaks left, remove the region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$remove_double_peaks_in_scans()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize_data"></a>}}
\if{latex}{\out{\hypertarget{method-normalize_data}{}}}
\subsection{Method \code{normalize_data()}}{
Normalize the intensity data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$normalize_data(which_data = "both")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_data}}{raw, characterized, or both (default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_peaks_in_regions"></a>}}
\if{latex}{\out{\hypertarget{method-find_peaks_in_regions}{}}}
\subsection{Method \code{find_peaks_in_regions()}}{
Find the peaks in the regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$find_peaks_in_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_mzsd"></a>}}
\if{latex}{\out{\hypertarget{method-model_mzsd}{}}}
\subsection{Method \code{model_mzsd()}}{
Model the m/z standard deviation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$model_mzsd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model_heightsd"></a>}}
\if{latex}{\out{\hypertarget{method-model_heightsd}{}}}
\subsection{Method \code{model_heightsd()}}{
Model the intensity height standard deviation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$model_heightsd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-indicate_high_frequency_sd"></a>}}
\if{latex}{\out{\hypertarget{method-indicate_high_frequency_sd}{}}}
\subsection{Method \code{indicate_high_frequency_sd()}}{
Look for peaks with higher than expected frequency standard deviation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$indicate_high_frequency_sd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_data}{}}}
\subsection{Method \code{add_data()}}{
Add the data from an SCMzml object to the underlying SCPeakRegions object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$add_data(sc_mzml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc_mzml}}{the SCMzml object being passed in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-summarize_peaks}{}}}
\subsection{Method \code{summarize_peaks()}}{
Summarize the peaks to go into JSON form.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$summarize_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_offset"></a>}}
\if{latex}{\out{\hypertarget{method-add_offset}{}}}
\subsection{Method \code{add_offset()}}{
Add an offset based on width in frequency space to m/z to describe how wide the peak is.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$add_offset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sort_ascending_mz"></a>}}
\if{latex}{\out{\hypertarget{method-sort_ascending_mz}{}}}
\subsection{Method \code{sort_ascending_mz()}}{
Sort the data in m/z order, as the default is frequency order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$sort_ascending_mz()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-characterize_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-characterize_peaks}{}}}
\subsection{Method \code{characterize_peaks()}}{
Run the overall peak characterization from start to finish.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$characterize_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summarize everything for output to the zip file after completion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$summarize(
  package_used = "package:ScanCentricPeakCharacterization"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{package_used}}{which package is being used for this work.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peak_meta"></a>}}
\if{latex}{\out{\hypertarget{method-peak_meta}{}}}
\subsection{Method \code{peak_meta()}}{
Generate the meta data that goes into the accompanying JSON file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$peak_meta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make a new SCPeakRegionFinder object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$new(
  sc_mzml = NULL,
  sliding_region_size = 10,
  sliding_region_delta = 1,
  tiled_region_size = 1,
  tiled_region_delta = 1,
  region_percentile = 0.99,
  offset_multiplier = 1,
  frequency_multiplier = 400,
  quantile_multiplier = 1.5,
  n_point_region = 2000,
  peak_method = "lm_weighted",
  min_points = 4,
  n_zero_tiles = 1,
  zero_normalization = FALSE,
  calculate_peak_area = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc_mzml}}{the SCMzml object to use (can be missing)}

\item{\code{sliding_region_size}}{how wide to make the sliding regions in data points}

\item{\code{sliding_region_delta}}{how far apart are the starting locations of the sliding regions}

\item{\code{tiled_region_size}}{how wide are the tiled regions}

\item{\code{tiled_region_delta}}{how far apart are the tiled reigons}

\item{\code{region_percentile}}{cumulative percentile cutoff to use}

\item{\code{offset_multiplier}}{what offset multiplier should be used}

\item{\code{frequency_multiplier}}{how much to multiply frequency points to interval ranges}

\item{\code{quantile_multiplier}}{how much to adjust the quantile cutoff by}

\item{\code{n_point_region}}{how many points in the large tiled regions}

\item{\code{peak_method}}{the peak characterization method to use (lm_weighted)}

\item{\code{min_points}}{how many points to say there is a peak (4)}

\item{\code{n_zero_tiles}}{how many tiles in a row do there need to be to split things up? (1)}

\item{\code{zero_normalization}}{don't actually do normalization (FALSE)}

\item{\code{calculate_peak_area}}{should peak area as well as peak height be returned? (FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegionFinder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
