% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{SCPeakRegions}
\alias{SCPeakRegions}
\title{Holds all the peak region data}
\description{
Holds all the peak region data

Holds all the peak region data
}
\details{
This reference class represents the peak region data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{frequency_point_regions}}{the frequency data}

\item{\code{frequency_fit_description}}{the model of frequency ~ m/z}

\item{\code{mz_fit_description}}{the model of m/z ~ frequency}

\item{\code{peak_regions}}{the peak regions}

\item{\code{sliding_regions}}{the sliding regions used for density calculations}

\item{\code{tiled_regions}}{the tiled regions used for grouping and splitting peak regions}

\item{\code{peak_region_list}}{list of regions}

\item{\code{frequency_multiplier}}{how much to multiplier frequency by to make interval points}

\item{\code{scan_peaks}}{the peaks by scans}

\item{\code{peak_data}}{the data.frame of final peak data}

\item{\code{scan_level_arrays}}{scan level peak data as matrices}

\item{\code{is_normalized}}{are the peak intensities normalized}

\item{\code{normalization_factors}}{the normalization factors calculated}

\item{\code{n_scan}}{how many scans are we working with}

\item{\code{scans_per_peak}}{??}

\item{\code{scan_perc}}{what percentage of scans is a minimum}

\item{\code{min_scan}}{based on \code{scan_perc}, how many scans minimum does a peak
need to be in}

\item{\code{max_subsets}}{??}

\item{\code{scan_subsets}}{??}

\item{\code{frequency_range}}{what is the range in frequency space}

\item{\code{scan_correlation}}{??}

\item{\code{keep_peaks}}{which peaks are we keeping out of all the peaks we had}

\item{\code{peak_index}}{the indices for the peaks}

\item{\code{scan_indices}}{the names of the scans}

\item{\code{instrument}}{the instrument serial number if available}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_min_scan}{\code{SCPeakRegions$set_min_scan()}}
\item \href{#method-add_data}{\code{SCPeakRegions$add_data()}}
\item \href{#method-new}{\code{SCPeakRegions$new()}}
\item \href{#method-clone}{\code{SCPeakRegions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_min_scan"></a>}}
\if{latex}{\out{\hypertarget{method-set_min_scan}{}}}
\subsection{Method \code{set_min_scan()}}{
sets the minimum number of scans to use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegions$set_min_scan()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_data}{}}}
\subsection{Method \code{add_data()}}{
Adds the data from an SCMzml object to the SCPeakRegion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegions$add_data(sc_mzml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc_mzml}}{the SCMzml object being passed in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new SCPeakRegions object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegions$new(
  sc_mzml = NULL,
  frequency_multiplier = 400,
  scan_perc = 0.1,
  max_subsets = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc_mzml}}{the SCMzml object to get data from}

\item{\code{frequency_multiplier}}{how much to multiply frequency by}

\item{\code{scan_perc}}{how many scans are required to be in to be a "peak"}

\item{\code{max_subsets}}{??}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCPeakRegions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
