% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\docType{data}
\name{SCZip}
\alias{SCZip}
\title{Represents the zip mass spec file}
\format{
An object of class \code{R6ClassGenerator} of length 25.
}
\usage{
SCZip
}
\value{
SCZip object
}
\description{
This reference class represents the zip mass spec file. It does this by
providing objects for the zip file, the metadata, as well as various bits
underneath such as the raw data and peak lists, and their
associated metadata. Although it is possible to work with the SCSCZip object directly, it
is heavily recommended to use the SCCharacterizePeaks object
for carrying out the various steps of an analysis, including peak finding.
}
\section{\code{SCZip} Methods}{


use \code{?method-name} to see more details about each individual method

\describe{
\item{\code{sc_zip}}{make a new \code{SCZip}}
\item{\code{show_temp_dir}}{show where files are stored}
\item{\code{save}}{save the file}
\item{\code{cleanup}}{unlink the \code{temp_directory}}
\item{\code{add_peak_list}}{add a \code{Peaks} to the data}
}
}

\section{\code{SCZip} Data Members}{

\describe{
\item{\code{zip_file}}{the zip file that was read in}
\item{\code{metadata}}{the actual metadata for the file}
\item{\code{metadata_file}}{the metadata file}
\item{\code{sc_raw}}{a \code{SCRaw} holding the raw data}
\item{\code{peaks}}{a \code{Peaks} holding the peak analysis}
\item{\code{id}}{the sample id}
\item{\code{out_file}}{the file where data will be saved}
}
}

\seealso{
SCCharacterizePeaks
}
\keyword{datasets}
