% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{set_internal_map}
\alias{set_internal_map}
\title{pick mapping function}
\usage{
set_internal_map(map_function = NULL)
}
\arguments{
\item{map_function}{which function to use, assigns it to an internal object}
}
\description{
Allows the user to set which mapping function is being used internally in the functions.
}
\details{
by default, the package uses {purrr::map} to iterate over things.
However, if you have the {furrr} package installed, you could switch it
to use {furrr::future_map} instead.
}
\examples{
\dontrun{
 library(furrr)
 future::plan(multicore)
 set_internal_map(furrr::future_map)
}

}
