% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{single_pass_normalization}
\alias{single_pass_normalization}
\title{Single Pass Normalization}
\usage{
single_pass_normalization(
  scan_peaks,
  intensity_measure = c("RawHeight", "Height"),
  summary_function = median,
  use_peaks = NULL,
  min_ratio = 0.7
)
}
\arguments{
\item{scan_peaks}{the scan peaks to normalize}

\item{intensity_measure}{which intensities to normalize}

\item{summary_function}{which function to use to calculate summaries (median)}

\item{use_peaks}{which peaks to use for normalization}

\item{min_ratio}{what ratio of maximum intensity of peaks should we use for normalization}
}
\value{
scan_peaks list
}
\description{
Does a single pass of normalizing scans to each other.
}
