% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCPeakRegions.R
\name{split_region_by_peaks}
\alias{split_region_by_peaks}
\title{split signal region}
\usage{
split_region_by_peaks(
  region_list,
  min_points = 4,
  metadata = NULL,
  calculate_peak_area = FALSE
)
}
\arguments{
\item{region_list}{a list with points and tiles IRanges objects}

\item{min_points}{how many points are needed for a peak}

\item{metadata}{metadata that tells how things should be processed}
}
\value{
list
}
\description{
Given a region that should contain signal, and the point data within it,
find the peaks, and return the region, and the set of points that make
up each peak from each scan.
}
