% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{ssr}
\alias{ssr}
\title{sum of squares residuals}
\usage{
ssr(object)
}
\arguments{
\item{object}{the lm object}
}
\value{
numeric
}
\description{
returns the sum of squares residuals from an \code{lm} object
}

