% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{subset_samples}
\alias{subset_samples}
\title{generate subset samples}
\usage{
subset_samples(n_indices, n_point = 3, max_subsets = 100,
  min_indices = 4)
}
\arguments{
\item{min_indices}{}
}
\value{
list of indices
}
\description{
From N \code{n_indices}, find the \code{n_point} combinations, up to \code{max_subsets} if
there are more than them available. See \strong{Details} for more information.
}
