% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakMS.R
\name{summarize_correspondencelist}
\alias{summarize_correspondencelist}
\title{summarize_correspondencelist}
\usage{
summarize_correspondencelist(correspondencelist_object,
  averaged_values = TRUE, individual_values = FALSE)
}
\arguments{
\item{correspondencelist_object}{the correspondent peaks to summarize}

\item{individual_values}{should individual samples / scans be returned}

\item{peakfinder_obj}{a PeakFinder that also needs to be summarized}

\item{average_values}{should values be averaged}

\item{package_used}{which package was used to do summarize}
}
\value{
list
}
\description{
creates a list of lists, where each list is named according to the json file
it should generate. One list will be the processing information, and then
their will either be a list for each scan (sample) or one list of the average
values
}
