% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{test_peak_area}
\alias{test_peak_area}
\title{test points area}
\usage{
test_peak_area(mz_peak, min_points = 5, min_area = 0.1)
}
\arguments{
\item{mz_peak}{the mz and intensity values defining the peak}

\item{min_points}{the minimum number of points to include in the peak}
}
\value{
numeric
}
\description{
given an mz data.frame, test whether the non-zero intensity points account for a
significant amount of area and are a certain number of points
}
\details{
returns a matrix of vectors, where the vector is:
\enumerate{
  \item start of points
  \item end of points
  \item area
}
}
