% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{test_peak_rsq}
\alias{test_peak_rsq}
\title{test points}
\usage{
test_peak_rsq(mz_peak, min_points = 5)
}
\arguments{
\item{mz_peak}{the mz and intensity values defining the peak}

\item{min_points}{the minimum number of points to include in the peak}
}
\value{
numeric
}
\description{
given an mz data.frame, test all possible sets of points for their fit
to a parabolic function.
}
\details{
returns a matrix of vectors, where the vector is:
\enumerate{
  \item start of points
  \item end of points
  \item number of points
  \item r-square fit for the set of points
}
}

