% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_normalization.R
\name{tic_normalization}
\alias{tic_normalization}
\title{tic normalization}
\usage{
tic_normalization(intensity_matrix, tic)
}
\arguments{
\item{intensity_matrix}{the intensity matrix}

\item{tic}{vector of tic's}
}
\value{
transformed intensity matrix
}
\description{
given an intensity matrix and vector of total intensities, normalize each
column of the intensity matrix by the total intensities. Column names of
the matrix and names of the vector should be identical.
}
