% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_methods.R
\name{zip_list_contents}
\alias{zip_list_contents}
\title{list zip file contents}
\usage{
zip_list_contents(zip_file)
}
\arguments{
\item{zip_file}{the zip file}
}
\description{
given a zip file, list the contents
}
