% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{zip_ms}
\alias{zip_ms}
\title{make a new ZipMS}
\usage{
zip_ms(
  in_file,
  mzml_meta_file = NULL,
  out_file = NULL,
  load_raw = TRUE,
  load_peak_list = TRUE
)
}
\arguments{
\item{in_file}{the file to use (either .zip or .mzML)}

\item{mzml_meta_file}{metadata file (.json)}

\item{out_file}{the file to save to at the end}

\item{load_raw}{logical to load the raw data}

\item{load_peak_list}{to load the peak list if it exists}
}
\value{
ZipMS
}
\description{
make a new ZipMS
}
