% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{add_peak_list}
\alias{add_peak_list}
\title{ZipMS - add_peak_list}
\usage{
ZipMS$add_peak_list(peak_list_data)
}
\arguments{
\item{peak_list_data}{a \code{\link[=Peaks]{Peaks()}} object}
}
\description{
adds a peak list to the \code{ZipMS}
}
