% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeMSPeaks.R
\name{CharacterizeMSPeaks}
\alias{CharacterizeMSPeaks}
\title{analyze ftms mass-spec data}
\description{
This class allows you to analyze mass spec data, and controls the execution
of reading in the mass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_file}{\code{CharacterizeMSPeaks$load_file()}}
\item \href{#method-find_peaks}{\code{CharacterizeMSPeaks$find_peaks()}}
\item \href{#method-summarize}{\code{CharacterizeMSPeaks$summarize()}}
\item \href{#method-save_peaks}{\code{CharacterizeMSPeaks$save_peaks()}}
\item \href{#method-write_zip}{\code{CharacterizeMSPeaks$write_zip()}}
\item \href{#method-set_peak_finder}{\code{CharacterizeMSPeaks$set_peak_finder()}}
\item \href{#method-filter_raw_scans}{\code{CharacterizeMSPeaks$filter_raw_scans()}}
\item \href{#method-run_all}{\code{CharacterizeMSPeaks$run_all()}}
\item \href{#method-prep_data}{\code{CharacterizeMSPeaks$prep_data()}}
\item \href{#method-add_regions}{\code{CharacterizeMSPeaks$add_regions()}}
\item \href{#method-run_splitting}{\code{CharacterizeMSPeaks$run_splitting()}}
\item \href{#method-new}{\code{CharacterizeMSPeaks$new()}}
\item \href{#method-clone}{\code{CharacterizeMSPeaks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-load_file}{}}}
\subsection{Method \code{load_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$load_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-find_peaks}{}}}
\subsection{Method \code{find_peaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$find_peaks(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-save_peaks}{}}}
\subsection{Method \code{save_peaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$save_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_zip"></a>}}
\if{latex}{\out{\hypertarget{method-write_zip}{}}}
\subsection{Method \code{write_zip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$write_zip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_peak_finder"></a>}}
\if{latex}{\out{\hypertarget{method-set_peak_finder}{}}}
\subsection{Method \code{set_peak_finder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$set_peak_finder(in_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_raw_scans"></a>}}
\if{latex}{\out{\hypertarget{method-filter_raw_scans}{}}}
\subsection{Method \code{filter_raw_scans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$filter_raw_scans()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$run_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prep_data"></a>}}
\if{latex}{\out{\hypertarget{method-prep_data}{}}}
\subsection{Method \code{prep_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$prep_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_regions"></a>}}
\if{latex}{\out{\hypertarget{method-add_regions}{}}}
\subsection{Method \code{add_regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$add_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_splitting"></a>}}
\if{latex}{\out{\hypertarget{method-run_splitting}{}}}
\subsection{Method \code{run_splitting()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$run_splitting()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$new(
  in_file,
  metadata_file = NULL,
  out_file = NULL,
  peak_finder = NULL,
  temp_loc = NULL,
  raw_scan_filter = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacterizeMSPeaks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
