% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRaw.R
\name{filter_remove_outlier_scans_default}
\alias{filter_remove_outlier_scans_default}
\title{default outlier scan function}
\usage{
filter_remove_outlier_scans_default(sc_raw)
}
\arguments{
\item{sc_raw}{the sc_raw object}
}
\value{
SCRaw
}
\description{
default outlier scan function
}
\details{
This is the default filtering and removing outliers function.
It is based on the Moseley groups normal samples and experience.
However, it does not reflect everyone's experience and needs.
We expect that others have different use cases and needs, and therefore
they should create their own function and use it appropriately.

Please examine this function and write your own as needed.
It should at the very least take a SCRaw object, work on the scan_info slot,
and then create a column with the name "keep" denoting which scans to keep.
}
