% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{find the peaks}
\usage{
find_peaks(mz_data, min_points = 4, n_peak = Inf, flat_cut = 0.98)
}
\arguments{
\item{mz_data}{a data.frame of mz and intensity}

\item{min_points}{minimum number of points to define a peak}

\item{n_peak}{maximum number of peaks to return}

\item{flat_cut}{minimum ratio of high points to say peak is "flat"}
}
\value{
list
}
\description{
Uses the \code{pracma::findpeaks} regular expression diff algorithm to find
possible peaks, then tests each one to fit to parabolic model on log-intensities,
returning the statistics about the peak.
}

