% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_peaks.R
\name{get_fitted_peak_info}
\alias{get_fitted_peak_info}
\title{parabolic fitted peak}
\usage{
get_fitted_peak_info(possible_peak, use_loc = "mz", w = NULL,
  addend = 1e-08)
}
\arguments{
\item{possible_peak}{data.frame of mz, intensity and log intensity}

\item{use_loc}{which field to use for locations, default is "mz"}

\item{w}{the weights to use for the points}
}
\value{
data.frame
}
\description{
given the peak, simply returns the model and residuals for the full peak
and the area
}
