% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRaw.R
\name{get_scan_info}
\alias{get_scan_info}
\title{get scan level info}
\usage{
get_scan_info(mzml_data)
}
\arguments{
\item{mzml_data}{the MSnbase mzml data object}
}
\value{
data.frame, see \strong{Details}
}
\description{
get scan level info
}
\details{
returns a data.frame with:
\itemize{
\item \code{scanIndex}: the indices of the scans
\item \code{scan}: the number of the scan by number. This will be used to name scans.
\item \code{polarity}: +1 or -1 depending on if the scan is positive or negative
\item \code{rtime}: the retention time or injection time of the scan for for direct-injection data
\item \code{tic}: the total intensity of the scan
\item \code{rtime_lag}: how long between this scan and previous scan
\item \code{rtime_lead}: how long between this scan and next scan
}

After running \code{predict_frequency()}, the following fields are added
from the information returned from frequency conversion:
\itemize{
\item \code{mad}: mean absolute deviation of residuals
\item \verb{frequency model coefficients}: the coefficients from the
fit frequency, named whatever you named them
\item \verb{mz model coefficients}: similar, but for the m/z model
}
}
