% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{import_mzML}
\alias{import_mzML}
\title{import mzML data}
\usage{
import_mzML(mz_file, ...)
}
\arguments{
\item{mz_file}{the mzML file to import}

\item{...}{other \code{xcms::getSpec} parameters}
}
\value{
tbl_df
}
\description{
function to import mzML data and generate the mean spectrum. \code{mz_file}
should be the \emph{full path} to the data.
}

