% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{import_raw_ms}
\alias{import_raw_ms}
\title{import raw mass spec data}
\usage{
import_raw_ms(raw_data, profstep = 0, includeMSn = TRUE)
}
\arguments{
\item{raw_data}{the raw mass spec file to import}

\item{profstep}{the profile step to use, should be 0 to make our lives easier}

\item{includeMSn}{whether to include MSn data, should be TRUE to get \emph{all data}.}

\item{...}{other xcmsRaw parameters}
}
\value{
xcmsRaw
}
\description{
function to import raw mass spec data in a way that provides what we need to work
with it. \code{raw_data} should be the \emph{full path} to the data.
}

