% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{plot_tic}
\alias{plot_tic}
\title{plot tic}
\usage{
plot_tic(xcms_raw, color_ms = TRUE, log_transform = TRUE)
}
\arguments{
\item{xcms_raw}{an \code{xcmsRaw} object (ideally from \code{import_mzML})}

\item{color_ms}{should scans be colored by their \emph{ms} level and type?}
}
\value{
ggplot
}
\description{
function to plot the total intensity chromatogram of the data, with information
about which scans are which
}
