% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{raw_peak_intensity}
\alias{raw_peak_intensity}
\title{raw peaks intensities}
\usage{
raw_peak_intensity(sc_raw, scanrange = sc_raw$scan_range)
}
\arguments{
\item{sc_raw}{an RawMS object}

\item{scanrange}{the range of scans to use, default is derived from sc_raw}
}
\value{
data.frame of mz and intensity
}
\description{
generate raw peaks from just averaging scans via xcms
}
