% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipMS.R
\name{raw_peaks}
\alias{raw_peaks}
\title{raw peaks}
\usage{
raw_peaks(raw_ms, scanrange = raw_ms$scan_range)
}
\arguments{
\item{raw_ms}{an RawMS object}

\item{scanrange}{the range of scans to use, default is derived from raw_ms}
}
\value{
numeric
}
\description{
generate raw peaks from just averaging scans via xcms
}
