% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCharacterizePeaks.R
\name{SCCharacterizePeaks}
\alias{SCCharacterizePeaks}
\title{R6 Class Controlling Peak Characterization}
\description{
Peak characterization control
}
\details{
Peak characterization associates data with the \code{SCZip},
\code{SCPeakRegionFinder}, and controls their execution.
}
\examples{
\dontrun{
  lipid_sample = system.file("extdata", "lipid_example.mzML", package = "ScanCentricPeakCharacterization")
  sc_char = SCCharacterizePeaks$new(lipid_sample)

  # prep data and check model
  library(ggplot2)
  library(patchwork)
  sc_char$load_file()
  sc_char$prepare_mzml_data()
  sc_char$check_frequency_model()

  # run characterization
  save_loc = "test.zip"
  sc_char = SCCharacterizePeaks$new(lipid_sample, out_file = save_loc)
  sc_char$run_all()
}
}
\seealso{
\code{\link[=SCPeakRegionFinder$summarize]{SCPeakRegionFinder$summarize()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{found_peaks}}{peaks found by a function}

\item{\code{id}}{a holder for the ID of the sample}

\item{\code{filter_remove_outlier_scans}}{function to be used for filtering scans, see \link{filter_remove_outlier_scans_default} as an example}

\item{\code{choose_single_frequency_model}}{function to be used to choose a single model for frequency conversion, see \link{choose_single_frequency_model_default} as an example}

\item{\code{frequency_fit_description}}{the model for conversion to frequency}

\item{\code{mz_fit_description}}{the model for converting back to m/z}

\item{\code{sc_peak_region_finder}}{the peak finder object}

\item{\code{sc_zip}}{the \code{SCZip} that represents the final file}

\item{\code{in_file}}{the input file}

\item{\code{metadata_file}}{the metadata file}

\item{\code{out_file}}{where everything gets saved}

\item{\code{temp_loc}}{where intermediates get saved}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_file}{\code{SCCharacterizePeaks$load_file()}}
\item \href{#method-prepare_mzml_data}{\code{SCCharacterizePeaks$prepare_mzml_data()}}
\item \href{#method-set_frequency_fit_description}{\code{SCCharacterizePeaks$set_frequency_fit_description()}}
\item \href{#method-set_mz_fit_description}{\code{SCCharacterizePeaks$set_mz_fit_description()}}
\item \href{#method-set_filter_remove_outlier_scans}{\code{SCCharacterizePeaks$set_filter_remove_outlier_scans()}}
\item \href{#method-set_choose_single_frequency_model}{\code{SCCharacterizePeaks$set_choose_single_frequency_model()}}
\item \href{#method-predict_frequency}{\code{SCCharacterizePeaks$predict_frequency()}}
\item \href{#method-check_frequency_model}{\code{SCCharacterizePeaks$check_frequency_model()}}
\item \href{#method-get_frequency_data}{\code{SCCharacterizePeaks$get_frequency_data()}}
\item \href{#method-scan_info}{\code{SCCharacterizePeaks$scan_info()}}
\item \href{#method-find_peaks}{\code{SCCharacterizePeaks$find_peaks()}}
\item \href{#method-summarize}{\code{SCCharacterizePeaks$summarize()}}
\item \href{#method-save_peaks}{\code{SCCharacterizePeaks$save_peaks()}}
\item \href{#method-write_zip}{\code{SCCharacterizePeaks$write_zip()}}
\item \href{#method-run_all}{\code{SCCharacterizePeaks$run_all()}}
\item \href{#method-prep_data}{\code{SCCharacterizePeaks$prep_data()}}
\item \href{#method-add_regions}{\code{SCCharacterizePeaks$add_regions()}}
\item \href{#method-run_splitting}{\code{SCCharacterizePeaks$run_splitting()}}
\item \href{#method-new}{\code{SCCharacterizePeaks$new()}}
\item \href{#method-clone}{\code{SCCharacterizePeaks$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_file"></a>}}
\if{latex}{\out{\hypertarget{method-load_file}{}}}
\subsection{Method \code{load_file()}}{
Loads the mzml data into the \code{SCZip}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$load_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_mzml_data"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_mzml_data}{}}}
\subsection{Method \code{prepare_mzml_data()}}{
Prepare the mzml data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$prepare_mzml_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_frequency_fit_description"></a>}}
\if{latex}{\out{\hypertarget{method-set_frequency_fit_description}{}}}
\subsection{Method \code{set_frequency_fit_description()}}{
Set the frequency fit description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$set_frequency_fit_description(frequency_fit_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{frequency_fit_description}}{the frequency model description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_mz_fit_description"></a>}}
\if{latex}{\out{\hypertarget{method-set_mz_fit_description}{}}}
\subsection{Method \code{set_mz_fit_description()}}{
Set the mz fit description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$set_mz_fit_description(mz_fit_description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mz_fit_description}}{the m/z model description}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_filter_remove_outlier_scans"></a>}}
\if{latex}{\out{\hypertarget{method-set_filter_remove_outlier_scans}{}}}
\subsection{Method \code{set_filter_remove_outlier_scans()}}{
Sets the scan filtering and check for outlier function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$set_filter_remove_outlier_scans(
  filter_remove_outlier_scans
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter_remove_outlier_scans}}{the function to remove outlier scans}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_choose_single_frequency_model"></a>}}
\if{latex}{\out{\hypertarget{method-set_choose_single_frequency_model}{}}}
\subsection{Method \code{set_choose_single_frequency_model()}}{
Sets the function for choosing a single frequency model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$set_choose_single_frequency_model(
  choose_single_frequency_model
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{choose_single_frequency_model}}{the function for choosing a single model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_frequency"></a>}}
\if{latex}{\out{\hypertarget{method-predict_frequency}{}}}
\subsection{Method \code{predict_frequency()}}{
Run frequency prediction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$predict_frequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_frequency_model"></a>}}
\if{latex}{\out{\hypertarget{method-check_frequency_model}{}}}
\subsection{Method \code{check_frequency_model()}}{
Check the frequency model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$check_frequency_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_frequency_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_frequency_data}{}}}
\subsection{Method \code{get_frequency_data()}}{
Get the frequency data from the \code{SCMzml} bits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$get_frequency_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scan_info"></a>}}
\if{latex}{\out{\hypertarget{method-scan_info}{}}}
\subsection{Method \code{scan_info()}}{
Get the \code{SCMzml$scan_info} out
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$scan_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-find_peaks}{}}}
\subsection{Method \code{find_peaks()}}{
Do the peak characterization without saving
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$find_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Generates the JSON output summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_peaks"></a>}}
\if{latex}{\out{\hypertarget{method-save_peaks}{}}}
\subsection{Method \code{save_peaks()}}{
Saves the peaks and JSON to the temp file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$save_peaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_zip"></a>}}
\if{latex}{\out{\hypertarget{method-write_zip}{}}}
\subsection{Method \code{write_zip()}}{
Write the zip file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$write_zip()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
Runs all of the pieces for peak characterization in order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$run_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prep_data"></a>}}
\if{latex}{\out{\hypertarget{method-prep_data}{}}}
\subsection{Method \code{prep_data()}}{
Loads and preps the data for characterization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$prep_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_regions"></a>}}
\if{latex}{\out{\hypertarget{method-add_regions}{}}}
\subsection{Method \code{add_regions()}}{
Adds initial regions for finding real peak containing regions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$add_regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_splitting"></a>}}
\if{latex}{\out{\hypertarget{method-run_splitting}{}}}
\subsection{Method \code{run_splitting()}}{
Does initial region splitting and peak finding in scans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$run_splitting()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{SCCharacterizePeaks} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$new(
  in_file,
  metadata_file = NULL,
  out_file = NULL,
  temp_loc = NULL,
  frequency_fit_description = NULL,
  mz_fit_description = NULL,
  filter_remove_outlier_scans = NULL,
  choose_single_frequency_model = NULL,
  sc_peak_region_finder = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_file}}{the mass spec data file to use (required)}

\item{\code{metadata_file}}{a json metadata file (optional)}

\item{\code{out_file}}{where to save the final zip container}

\item{\code{temp_loc}}{a specified temporary location}

\item{\code{frequency_fit_description}}{mz -> frequency model}

\item{\code{mz_fit_description}}{frequency -> mz model}

\item{\code{filter_remove_outlier_scans}}{function for scan filtering}

\item{\code{choose_single_frequency_model}}{function to choose a single frequency model}

\item{\code{sc_peak_region_finder}}{a blank \code{SCPeakRegionFinder} to use instead of the default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCCharacterizePeaks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
