% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMzml.R
\name{SCMzml}
\alias{SCMzml}
\title{R6 Class For mzML Data}
\description{
mzML mass spectrometry data container with some useful methods.
}
\details{
Provides our own container for mzML data, and does conversion to frequency,
filtering scans, choosing a single frequency regression model, and
generating the frequency data for use in the peak characterization.
}
\examples{
\dontrun{
  lipid_sample = system.file("extdata", "lipid_example.mzML",
  package = "ScanCentricPeakCharacterization")
}
}
\seealso{
\code{\link[=filter_remove_outlier_scans_default]{filter_remove_outlier_scans_default()}}

\code{\link[=choose_single_frequency_model_default]{choose_single_frequency_model_default()}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mzml_metadata}}{metadata from an external json file}

\item{\code{mzml_data}}{the actual mzml data from MSnbase}

\item{\code{mzml_df_data}}{a list of data.frames of the data}

\item{\code{scan_range}}{the range of scans to be used}

\item{\code{rtime_range}}{the range of retention times to keep}

\item{\code{mz_range}}{the mz range to use}

\item{\code{scan_info}}{data.frame of scan information}

\item{\code{remove_zero}}{should zero intensity data points be removed?}

\item{\code{frequency_fit_description}}{the model for m/z -> frequency}

\item{\code{mz_fit_description}}{the model for going from frequency -> m/z}

\item{\code{frequency_coefficients}}{the coefficients for the frequency model}

\item{\code{mz_coefficients}}{the coefficients for the m/z model}

\item{\code{difference_range}}{how wide to consider adjacent frequency points as \emph{good}}

\item{\code{filter_remove_outlier_scans}}{the function to remove scans that shouldn't be used as well as outliers}

\item{\code{choose_single_frequency_model}}{function to choose a single frequency model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-extract_mzml_data}{\code{SCMzml$extract_mzml_data()}}
\item \href{#method-predict_frequency}{\code{SCMzml$predict_frequency()}}
\item \href{#method-convert_to_frequency}{\code{SCMzml$convert_to_frequency()}}
\item \href{#method-check_frequency_model}{\code{SCMzml$check_frequency_model()}}
\item \href{#method-get_instrument}{\code{SCMzml$get_instrument()}}
\item \href{#method-get_frequency_data}{\code{SCMzml$get_frequency_data()}}
\item \href{#method-new}{\code{SCMzml$new()}}
\item \href{#method-clone}{\code{SCMzml$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_mzml_data"></a>}}
\if{latex}{\out{\hypertarget{method-extract_mzml_data}{}}}
\subsection{Method \code{extract_mzml_data()}}{
get the mzml data into data.frame form so we can use it
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$extract_mzml_data(remove_zero = self$remove_zero)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remove_zero}}{whether to remove zero intensity points or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_frequency"></a>}}
\if{latex}{\out{\hypertarget{method-predict_frequency}{}}}
\subsection{Method \code{predict_frequency()}}{
predict frequency and generate some summary information.
This does regression of frequency ~ m/z for each scan separately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$predict_frequency(
  frequency_fit_description = self$frequency_fit_description,
  mz_fit_description = self$mz_fit_description
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{frequency_fit_description}}{the regression model definition}

\item{\code{mz_fit_description}}{the regression model definition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_to_frequency"></a>}}
\if{latex}{\out{\hypertarget{method-convert_to_frequency}{}}}
\subsection{Method \code{convert_to_frequency()}}{
actually do the conversion of m/z to frequency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$convert_to_frequency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_frequency_model"></a>}}
\if{latex}{\out{\hypertarget{method-check_frequency_model}{}}}
\subsection{Method \code{check_frequency_model()}}{
check how well a given frequency model works for this data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$check_frequency_model(scan = 1, as_list = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scan}}{which scan to show predictions for}

\item{\code{as_list}}{whether plots should be returned as a single plot or a list of plots}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_instrument"></a>}}
\if{latex}{\out{\hypertarget{method-get_instrument}{}}}
\subsection{Method \code{get_instrument()}}{
get instrument data from associated mzml file metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$get_instrument()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_frequency_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_frequency_data}{}}}
\subsection{Method \code{get_frequency_data()}}{
get the frequency data to go into the next steps of analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$get_frequency_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$new(
  mzml_file,
  frequency_fit_description = c(a.freq = 0, x.freq = -1, y.freq = -1/2, z.freq = -1/3),
  mz_fit_description = c(a.mz = 0, x.mz = -1, y.mz = -2, z.mz = -3),
  metadata_file = NULL,
  scan_range = NULL,
  rtime_range = NULL,
  mz_range = NULL,
  remove_zero = FALSE,
  filter_remove_outlier_scans = generate_scan_outlier_filter(rtime = NA, y.freq = NA),
  choose_single_frequency_model = choose_single_frequency_model_default
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mzml_file}}{the file to load and use}

\item{\code{frequency_fit_description}}{a description of the regression model for frequency ~ m/z}

\item{\code{mz_fit_description}}{a description of the regression model for m/z ~ frequency}

\item{\code{metadata_file}}{a metadata file generated by ...}

\item{\code{scan_range}}{which scans can be used for analysis}

\item{\code{rtime_range}}{the retention time to use for scans}

\item{\code{mz_range}}{what m/z range to use}

\item{\code{remove_zero}}{should zero intensity data be removed?}

\item{\code{filter_remove_outlier_scans}}{the function to use to filter scans}

\item{\code{choose_single_frequency_model}}{the function to use to choose single frequency regression model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCMzml$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
