% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCZip.R
\name{SCZip}
\alias{SCZip}
\title{Represents the zip mass spec file}
\description{
Represents the zip mass spec file

Represents the zip mass spec file
}
\details{
This reference class represents the zip mass spec file. It does this by
providing objects for the zip file, the metadata, as well as various bits
underneath such as the mzml data and peak lists, and their
associated metadata. Although it is possible to work with the SCZip object directly, it
is heavily recommended to use the SCCharacterizePeaks object
for carrying out the various steps of an analysis, including peak finding.
}
\seealso{
\link{SCMzml}

\link{SCPeakRegionFinder}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{zip_file}}{the actual zip file}

\item{\code{zip_metadata}}{the metadata about the zip file}

\item{\code{metadata}}{the metadata itself}

\item{\code{metadata_file}}{the metadata file}

\item{\code{sc_mzml}}{the mzML data object.}

\item{\code{peaks}}{??}

\item{\code{sc_peak_region_finder}}{the peak finder object}

\item{\code{json_summary}}{jsonized summary of the peak characterization}

\item{\code{id}}{the identifier of the sample}

\item{\code{out_file}}{where to put the final file}

\item{\code{temp_directory}}{where we keep everything until peak characterization is done}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_mzml}{\code{SCZip$load_mzml()}}
\item \href{#method-load_sc_peak_region_finder}{\code{SCZip$load_sc_peak_region_finder()}}
\item \href{#method-save_json}{\code{SCZip$save_json()}}
\item \href{#method-save_sc_peak_region_finder}{\code{SCZip$save_sc_peak_region_finder()}}
\item \href{#method-load_peak_list}{\code{SCZip$load_peak_list()}}
\item \href{#method-compare_mzml_corresponded_densities}{\code{SCZip$compare_mzml_corresponded_densities()}}
\item \href{#method-new}{\code{SCZip$new()}}
\item \href{#method-show_temp_dir}{\code{SCZip$show_temp_dir()}}
\item \href{#method-write_zip}{\code{SCZip$write_zip()}}
\item \href{#method-cleanup}{\code{SCZip$cleanup()}}
\item \href{#method-finalize}{\code{SCZip$finalize()}}
\item \href{#method-add_peak_list}{\code{SCZip$add_peak_list()}}
\item \href{#method-clone}{\code{SCZip$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_mzml"></a>}}
\if{latex}{\out{\hypertarget{method-load_mzml}{}}}
\subsection{Method \code{load_mzml()}}{
Loads the mzML file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$load_mzml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_sc_peak_region_finder"></a>}}
\if{latex}{\out{\hypertarget{method-load_sc_peak_region_finder}{}}}
\subsection{Method \code{load_sc_peak_region_finder()}}{
Loads the SCPeakRegionFinder object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$load_sc_peak_region_finder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_json"></a>}}
\if{latex}{\out{\hypertarget{method-save_json}{}}}
\subsection{Method \code{save_json()}}{
Save the jsonized summary out to actual json files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$save_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_sc_peak_region_finder"></a>}}
\if{latex}{\out{\hypertarget{method-save_sc_peak_region_finder}{}}}
\subsection{Method \code{save_sc_peak_region_finder()}}{
Saves the SCPeakRegionFinder binary object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$save_sc_peak_region_finder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_peak_list"></a>}}
\if{latex}{\out{\hypertarget{method-load_peak_list}{}}}
\subsection{Method \code{load_peak_list()}}{
loads just the peak list data-frame instead of peak region finder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$load_peak_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare_mzml_corresponded_densities"></a>}}
\if{latex}{\out{\hypertarget{method-compare_mzml_corresponded_densities}{}}}
\subsection{Method \code{compare_mzml_corresponded_densities()}}{
compare peak densities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$compare_mzml_corresponded_densities(
  mz_range = c(150, 1600),
  window = 1,
  delta = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mz_range}}{the mz range to work over}

\item{\code{window}}{the window size in m/z}

\item{\code{delta}}{how much to move the window}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SCZip object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$new(
  in_file,
  mzml_meta_file = NULL,
  out_file = NULL,
  load_mzml = TRUE,
  load_peak_list = TRUE,
  temp_loc = tempfile("zipms")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_file}}{the mzML file to load}

\item{\code{mzml_meta_file}}{an optional metadata file}

\item{\code{out_file}}{where to save the final file}

\item{\code{load_mzml}}{should the mzML file actually be loaded into an SCMzml object?}

\item{\code{load_peak_list}}{should the peak list be loaded if this is previously characterized?}

\item{\code{temp_loc}}{where to make the temp file while working with the data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_temp_dir"></a>}}
\if{latex}{\out{\hypertarget{method-show_temp_dir}{}}}
\subsection{Method \code{show_temp_dir()}}{
Show the temp directory where everything is being worked with
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$show_temp_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_zip"></a>}}
\if{latex}{\out{\hypertarget{method-write_zip}{}}}
\subsection{Method \code{write_zip()}}{
Write the zip file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$write_zip(out_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_file}}{where to save the zip file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanup"></a>}}
\if{latex}{\out{\hypertarget{method-cleanup}{}}}
\subsection{Method \code{cleanup()}}{
delete the temp directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$cleanup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
delete when things are done
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_peak_list"></a>}}
\if{latex}{\out{\hypertarget{method-add_peak_list}{}}}
\subsection{Method \code{add_peak_list()}}{
Add peak list data to the temp directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$add_peak_list(peak_list_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peak_list_data}}{the peak list data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SCZip$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
