% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakRegions.R
\name{split_region_by_peaks}
\alias{split_region_by_peaks}
\title{split signal region}
\usage{
split_region_by_peaks(frequency_point_regions, tiled_regions,
  peak_method = "lm_weighted", min_points = 4)
}
\arguments{
\item{frequency_point_regions}{the frequency point regions to use}

\item{tiled_regions}{the tiled regions}

\item{peak_method}{the method for getting the peaks}

\item{min_points}{how many points are needed for a peak}
}
\value{
list
}
\description{
Given a region that should contain signal, and the point data within it,
find the peaks, and return the region, and the set of points that make
up each point from each scan.
}
